[roll] Roll fuchsia [component] Outgoing directory escrow support
This CL wires in the escrow actor into the start and stop actions.
When starting a component, the escrowed state is reaped and passed to
the program. When stopping a component, the escrowed state is stored.

Since the actor starts the component when the outgoing dir is readable,
calling `open_outgoing` will also start the component. Thus the
INSTANCE_NOT_RUNNING RealmQuery error is deprecated, and
INSTANCE_FAILED_TO_START is added.

Original-Bug: 319754472
Run-All-Tests: True

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1000121
Original-Revision: bd04154c0539dce25a576a37e4d7a18a41735d29
GitOrigin-RevId: d2df1252bdfaf0c4096ca88305d0fd5c6948c106
Change-Id: Ia177d6aa2eeaecc24837febdca537290262f3400