[roll] Roll fuchsia [fxfs] Switch BestFit to use BTreeMap.
I'm planning on capping the number of free extents kept per size bucket.
This change is in preparation for that but it also brings potential
minor memory savings for now as a side effect.

We (will) have a finite number of size buckets for allocations
based on number of blocks. Rather than store two 64-bit integers for
start/end of each range, this change allows us to store a set of start
offsets for each size bucket. Even with the overheads of BTreeSet, this
should lead to memory savings at scale.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014892
Original-Revision: 036acdfa1a8459dae85063de5f38e57827a1abfe
GitOrigin-RevId: 3a17aaaca957dc952b7d482d56f44601ee664959
Change-Id: I7e0dd57b7344fa76e5cdc6b106d6776613cbb398