[roll] Roll fuchsia [test][power] Add a local power monitor capture script.
Add a local python script which can be used to capture data from a
Monsoon power monitor.  The script is compatible with the current
mock-power-monitor capture script, however this one will actually talk
to a locally connected Monsoon and collect data.

It will account for the three main causes of clock drift which were
showing up in captured trace data.

1) Samples silently dropped by the capture script.
2) Calibration samples silently stripped form the data stream.
3) Monsoons fundamentally missing their nominal capture rate by many
   hundreds of ppm.

Additionally, it provide some options for how to handle dropped and
calibration samples, and for capturing the monitor's aux channel.
These features are not currently being used by the test framework,
but they could be leveraged in the future to provide better ways of
doing ongoing clock synchronization, and different ways to visualize
dropped or calibration samples.

Note that this script is useful for individual developers who happen
to have a Monsoon at their desk and are doing power investigations.
Automated testing infrastructure will need additional work in order to
integrate this script (or something like it) with their system.

Original-Bug: 327634476

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005932
Original-Revision: ad970f8f14d0035865ac35a85bab0d37247cafaf
GitOrigin-RevId: 5dc339a3a77403b4547d1f7ce98ae64374f12766
Change-Id: I290714924ac41d7b2e865ca2394729e218c7ecb5