[roll] Roll fuchsia [test][intl] Fixes change detector tests for date l10n
The about-to-be-rolled version of the ICU library changes the way
localized times are formatted for Spanish. It turns out that
such changes happen all the time, which makes this test brittle
in general.

And with a diminished interest in l10n on Fuchsia, I am not quite
sure how useful this example test is.

Let's patch it up now (by replacing the changed part of the string
with a permissive regex), and if it continues to generate churn,
consider removing it completely.

Original-Bug: b/330351469
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014633
Original-Revision: 17b5b0dc308909c2316461c8b64ca4ae5d114cac
GitOrigin-RevId: ff274c405094850b0a44ee0c870ec18544df1ef5
Change-Id: I31ff459ffdda0ab2b4d2374a8d537a1bb0a60625