[roll] Roll fuchsia [vkext] Use config to choose what tests to run
Use configs to determine what GPU to use and whether to test for several
of features. This will allow us to set the configuration OOT when this
test is used as a subpackage. This also allows us to recombine
vkext_protected and vkext_unprotected binaries, since we can disable a
flag to run only the unprotected tests from the binary.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014198
Original-Revision: 5386355614d4da13e3e6a0088631be899742d1f3
GitOrigin-RevId: 989c6321aa028dafc45080ede4f1c571f22e1492
Change-Id: Ic5a84e42c6a7512b395d044995e7b2134e3532ba