[roll] Roll fuchsia [ffx] More flexible Error handling
When deciding whether an `Error` is an `FfxError` or not in order
to convert it into `command` error, first check whether it is
_already_ a command error.

This may reduce the occurrence of some `BUG` messages for user errors
(or alternatively, make it easier to generate user errors).

The error layering code in `ffx` is needlessly complex, so this is
really a hack to avoid dealing with some of the confusing code. But
it shouldn't harm anything.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1012792
Original-Revision: 022c258395db40982b6379e44d9d3d7f0c04e9bd
GitOrigin-RevId: f8a2944e99cd0c75101b8928d72955b9b37fb2a7
Change-Id: Id047d52bdb475f95dc6bc4d49ad29ed20e69b663