[roll] Roll fuchsia [fx][ssh] Moving sshconfig generation to tools/devshell
This moves the generation of the sshconfig file used
by `fx shell` to be done at runtime vs. as part of the
build.

This is done to enable using ffx configuration to manage the
ssh key path both in and out of tree in order to avoid mismatched
keys or confusion on which key is used when. Since the configuration
can change outside the build, it makes more sense to check it at runtime
vs. build time.

More info: go/fuchsia-ssh-unification

Original-Bug: b/329486457

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001693
Original-Revision: 0182a8105356ee714f8609bc1db8deb733f92c70
GitOrigin-RevId: c244e8e110765756f9388df2d3102a3466092b72
Change-Id: Id247821b9932423d49e482fa32ddac371c858119