[roll] Roll fuchsia [fxtest] Fix positional args with --test-filter
argparse does not like interleaving positional args with switches, and
when we use the correct method (parse_intermixed_arguments) ordering of
switches is not preserved.

This is a problem for our selection strategy, which was previously to
allow switches like --and or --component to accept nargs="*", which
would then all get merged in order into the selection attribute of the
namespace (it didn't matter if your args were part of --and or the
overall "selection" positional args, it would all go to the same place).

This, however, does not work with args like --test-filter, which expect
nargs="*" and would use each arg for a new purpose. This means the
following breaks:
$ fx test foo --test-filter bar* baz

The intent is to filter for bar* is foo or baz, but the result is filter
bar* or baz in foo!

This CL fixes the problem as follows:
- selection is the only positional args with allows nargs="*"
- --and, --component, and --package have special handling in
  SelectionAction, where those flags are pre-processed to appear as
  positional args to argparse, allowing SelectionAction to recover the
  intended flags and ordering.
- To prevent bugs, those special handled args have a new InvalidAction
  which raises an runtime error if they ever make it through
  preprocessing untouched.
- --test-filter has been changed to support only nargs=1 with an append
  action, so it must be specified before each filter as intended.

Original-Fixed: 329894664

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014461
Original-Revision: 85a7fa3a674c928f78c8de6aefb9b6389a6b4023
GitOrigin-RevId: 0ed7ec779b25793d626d311dca714f7394872ed5
Change-Id: I850a745306e3484f1eddc1d184b650b989734a7f