[roll] Roll fuchsia [console][assembly] Move serial console to AIB
While this AIB is currently always included, we can replace the kernel
commandline parameter where we disable the shell with avoiding including
this AIB instead in the future. There are a number of use cases for the
kernel commandline parameter which make doing that difficult today.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013095
Original-Revision: 708ffeffae619eec8e25c8850a9b8233330be90e
GitOrigin-RevId: 9581c94354c02c3bb2ec426fa7bef0051f4a4e33
Change-Id: Ia3938346fce8c55279a862781b71a6d3b41e0ceb