[roll] Roll fuchsia [kernel][sched] Update RunQueue tiebreaker policy
This change updates (and officially documents) RunQueue's tiebreaker
story. This is made to conform to what we have in place in the kernel
today.

The policy now: the next thread is the one that is active and has the
earliest finish time within its current period. In the event of a tie of
finish times, the one with the earlier start time is picked - and in
then in the event of a tie of start times, the thread with the lowest
address is expediently picked (which is a small bias that should not
persist across runs).

Original-Bug: 328641440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015555
Original-Revision: e473993b77d43fdc1c394819217dd332889f2d72
GitOrigin-RevId: b9edb36f59cb982875690136d6a6ab5e07d07571
Change-Id: Ib9d175e0093fce17d5a973a42b5b11b0ba3ef7c5