[roll] Roll fuchsia [tee] Runtime binary and TA loader
This adds a binary to host the runtime support and load Trusted
Applications from dynamic libraries. The runtime binary exports TEE_...
symbols to the global symbol table for the TA to resolve when it loads.
The runtime binary expects to be included in a package that includes a
ta_name file from which it can infer the name of the TA shared library.
In this change, the binary doesn't do anything with the TA once it is
loaded.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015693
Original-Revision: 3ab8dc53b6c2c13204fd4752c989f4042a808683
GitOrigin-RevId: 291bbc231cef9c9dcd3c3bd04bd582ec468c6b8c
Change-Id: I9ec86ac3ab6982406e013b76955a23677c9759e1