[roll] Roll fuchsia [kernel][sched] Rename "capacity" as "firm capacity"
This change eases the full introduction of the concepts of firm and
flexible capacities, the latter of which will be introduced in a
follow-on change. That introduction will also bring with it a new
"flexible" bandwidth parameter, so it also seemed sensible to reverse
the ordering of sched::BandwidthParameter fields as first period and
then firm capacity. (So we end up with (period, firm, flexible) and not
(firm, period, flexible)).

Original-Bug: 328641440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015695
Original-Revision: 391afe76934fcb428a276343cc5577fe9f274008
GitOrigin-RevId: e50fd661fd4638222efda9e9733b04b4463b771f
Change-Id: Iff96bb20c91b8bdfe4ac0aeb717e4e5fd50ced0e