[roll] Roll fuchsia [ufs] Support UFS power mode
This CL uses the STRAT STOP UNIT SCSI command and DME HIBERNATE UIC
command to change the power mode of a UFS device.
Since the suspend/resume of the device driver framework is not complete
yet, we verify it with a unit test.

Test: fx test ufs-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939572
Original-Revision: 370894ba106cedf5f9940219b961796ebf541cae
GitOrigin-RevId: b15341a1cd8dcdc8fb8d6770471f63df6823b884
Change-Id: I63a2156274dce27d0e6d0c9e4702f819ffa4b6a9