[roll] Roll fuchsia [fdf] Don't destroy node on bind fail
Previously the node would be removed if
a driver failed to bind to it.  This was
an incorrect behavior that was not caught
in DFv2.

The test for this is the bind-fail-test,
which will check for the node existing when
we add the controller Allowlist.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015134
Original-Revision: bebd9be95fa40e5aa06383ed979fb1e7f6738141
GitOrigin-RevId: f8786774bbe45a6244fa6d7f1151347d0360642e
Change-Id: Ia044151737f343a1f06c39e4d601ee03e11c6e86