[roll] Roll fuchsia [component] delivery: on_readable support in the framework
When crafting the sender in make_program_outgoing_router, we read the
delivery field and tell the sender to wait for readable if delivery is
`on_readable`.

The waiting is performed in a scope that is shutdown when unresolved, in
line with other places where we clear escrowed data on unresolve to
prevent cross-version ABI concerns.

Adds an integration test of a component using `delivery: on_readable`,
where it should not receive requests unless the client writes to the
server_end.

There's an unforeseen interaction between on_readable and the
CapabilityRequested hook. Because the CapabilityRequested hook overrides
the capability obtained from the outgoing directory, we also lose any
laziness property. Properly handling that is split to the next CL for
ease of review, where we'll also add tests re CapabilityRequested.

Original-Bug: b/319754472

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015132
Original-Revision: 8b5b4c66653f183e20b8c2603180b89c568a02fc
GitOrigin-RevId: c5898acb57525363fa13397c259649841940ee94
Change-Id: Icd88724677dcc4460ede58578a45e4a74c368cda