[roll] Roll fuchsia [ffx][agis] Remove from being built-in
Removes ffx-agis from being built in to the main ffx binary, allowing it to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes when `ffx agis` is invoked, but makes it so that both the tool and ffx can build faster and with fewer resources making it easier to iterate on them.

Original-Bug: b/329326263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003649
Original-Revision: ad9c4d581a0af66e57c3ebcf760b6827bdd6b573
GitOrigin-RevId: b69904356b38e10c192a0182987e142556c84564
Change-Id: I67cb50c13700e513577369306e802fcd0de7d04a