[roll] Roll fuchsia [netstack3] Use IpDeviceSendContext in IGMP/MLD
The IGMP and MLD implementations currently depend on SendFrameContext,
which allows sending any kind of frame from a device, and the
implementation of it for IGMP and MLD specifically was, in practice,
delegating its implementation to IpDeviceSendContext, which is
specifically for sending IP packets.

Instead, depend on IpDeviceSendContext explicitly; this indicates in the
type system that the packets originating from IGMP/MLD are IP packets,
which will allow integrating the EGRESS IP layer filtering hook in the
implementation of IpDeviceSendContext more easily.

Test: netstack3-core-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014196
Original-Revision: f8a5f9980b5f2362a70888ba2979aee03162957e
GitOrigin-RevId: 1faa311ead5c92fb2c312f03f790f56bdc98fa67
Change-Id: I134aed2330213848f94d489fed53c1a11390d33f