[roll] Roll fuchsia [ffx][playground] Fix some weird grammar gotchas
We have ranges in playground, which meant "cd .." would parse the ".."
as an open range. That's uncomfortable (especially since open ranges
don't have any semantics yet), so we'll add the slightly arbitrary rule
that we always choose to parse as a bare string if we see
a dot at the start of the argument. You could still pass a range with
"cd (..)" if you needed to.

Also, integers don't clear their surroundings when they parse, so
"foocmd 123abc" parses like "foocmd 123 abc" with the 123 appearing in
the parse tree as an integer. We'll be a bit more forceful about the
whitespace at the end of the string to fix that.

Original-Fixed: b/325486174
Original-Fixed: b/325989896
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995237
Original-Revision: 6618e5af1973de6fa5ff787e6e2cab9d90cf4f0d
GitOrigin-RevId: abcb86d18dd43f941329ed4d9ac645a29db3e156
Change-Id: Ice775fc0e1cfe8dbe57f583a1405f53927b3e8e2