[roll] Roll fuchsia [ffx][repository] Auto-pick target if unset
This implements partial auto-selection of targets: If no default target
is explicitly set, the repo server will attempt to auto-detect the
target and, if there is only one, serve packages to this target. If
multiple targets are detected, the repo server does not guess. In this
case, the user is required to explicitly set the target.

Original-Fixed: b/329202124
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017034
Original-Revision: f5e1f6357b1d606fea8eb5b525b17eee65258f64
GitOrigin-RevId: d4a8729f81fd9acff31aeacd1d4ece6ea3374c03
Change-Id: I95073103c6e9e1a655578ba5189f865c33b91c81