[roll] Roll fuchsia [fx][build] Fix to support more than one target.
A recent CL broke the ability to pass multiple targets to
`fx build`, as only the first one was taken. The root cause
is that `read` can only read one line at a time, and must
be called in a loop.

Define a command-stdout-to-array function that uses `read`
in a loop with Bash 3.x (for MacOS), and the much more
efficient `readarray -t` for older Bash versions (Linux).

Original-Bug: 42084668
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017033
Original-Revision: 31e36ee816843510bdd06efd654cc88decec34d6
GitOrigin-RevId: 36afd6c7e76991243e0180f7917e374ff80dfb07
Change-Id: I131b31863f546a6db9da1334fbba7e406e850344