[roll] Roll fuchsia [debugger] Parallelize loading .dwo files.
LLVM can't handle accessing a binary on more than one thread, but with
debug fission compilation we have multiple binaries that can be loaded
in parallel. This improves indexing performance, taking about 60% of the
original time to load Chrome.

Adds a simple WorkerPool implementation from GN.

Adds a new file to manage the indexing to keep the threading logic out
of the already-complex ModuleSymbolsImpl.

Converts the Index member variable to a pointer so it can be moved from
the indexer to ModuleSymbolsImpl.

TEST=existing fission tests (there should be no behavior change).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016178
Original-Revision: e88a74ac80362a9565b1da96856fb4842486fdf4
GitOrigin-RevId: 4071ef26fb300d9cafa0841346c8572926c8104d
Change-Id: I89265ba486ee1a4dd22d6c69a9d3ff35e822d42c