[roll] Roll infra/recipes [checkout] Fix check for is_manifest_commit.
The check only checked that the commit project was the same as the
remote project, but if the host is different, then the checkout would
fail because it's looking for the commit id of a different repo. Instead
we should just compare the remote and commit_remote. This only affects
builds without a triggering commit or gerrit change (like led jobs)
which rely on recipe_wrapper to populate the gitiles commit with a
revision from the default public integration repo, which wouldn't work
for internal builders.

Original-Bug: 42073261
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1017160
Original-Revision: a4bf90bef0798d66199e36058463e799afd99edf
GitOrigin-RevId: 4dd72537e216c53ec4b7155a10d6f1d651a200c7
Change-Id: Ib1f4a233c440c1f8edbc403aba9a75fd41554e40