[roll] Roll fuchsia [trace_processing] Add all threads to model.processes
Adds a map to store the pid->tid mapping while importing a trace.
This is used to then populate the TraceModel's process.threads field so that
all threads, including threads that don't have events, are saved.
This is needed so that later we can we can map the thread to a name, and to
its process, when we print out CPU metrics.

Also sets a default thread name if there isn't one.

Test using `fx test --host`

Original-Bug: b/309000365, b/331457527

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1001908
Original-Revision: a14e8c872eecfebbf32e58117f60ffdf2fc86e77
GitOrigin-RevId: 729cf35922b297fdec7ebe1ad1d8d06042b0e2a8
Change-Id: I9599069a7bb26c5643ade4b2d49f8e1124bab48e