[roll] Roll fuchsia [testrunner] Return reconnect failures as connection failures.
For the SSHTester, we check in the beginning of a test if the SSH client
is still connected in case it got disconnected during the previous test.
However, if reconnecting fails, we'd just return the error like a normal
test error. This should get returned as a connection error so that
testrunner's logic will try to reconnect and if it fails, it will exit
early instead of continuing to run tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1010752
Original-Revision: 0c08559f65626a0fcea109891d040b39a98dce27
GitOrigin-RevId: 03a56d4cbf287976a860ec200e08abab0941c3cb
Change-Id: Ib732a0cab2e3e71a8482a1611d9773e04ab97847