[roll] Roll fuchsia [bt][rfcomm] Correctly set C/R bit in the Address Field
The Address Field of an RFCOMM packet contains a Command (C) / Response
(R) bit that signifies the type of packet being sent. The way it is set
depends on 1) The role of the device 2) Type of frame and 3) State of
the multiplexer.

Currently, the RFCOMM library does not use the frame type to make this
classification and instead uses the logic in GSM 5.2.1.2 to calculate
the C/R bit for all frame types. This is incorrect.

Update the lib to correctly set the C/R bit specifically when the type
of packet is a multiplexer command (UIH) by applying the classification
logic defined in GSM 5.4.3.1, which is a subclause of GSM 5.2.1.2.

Original-Bug: b/330690839
Test: fx test bt-rfcomm-tests
Test: fx test bt-rfcomm-lib-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009194
Original-Revision: 4226aa2fd401f6c09223f01bdaaacf5f58f3a93d
GitOrigin-RevId: 90bffe622c2d10d5797b4f3adfb496824a9f4bb4
Change-Id: I48053fdd37630a49a1f7a52e10992801006e97bb