[roll] Roll fuchsia [shush] Avoid parsing files with c-string literals
c-string literals were stabilized in rust 1.77, and are in use in
fuchsia source code. Unfortunately, proc-macro support was not
added for c-string literals, so syn is unable to process these
leading to a crash in syn if we try to parse any files with
c-string literals. Skip over any files and emit an error requesting
the affected lints to be manually updated in those files.

Original-Bug: 331979452
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017660
Original-Revision: 1d0d8bafcb25e61c426132706d71d112c4a39b45
GitOrigin-RevId: bbf9d0dc675bbbb859ed1558a4985e499daad103
Change-Id: Ic21403485c24521e5a5c4d36909b23ebb635d902