[roll] Roll fuchsia [profiler] Properly Clean Up Launched Components
If we launched a component that failed to resolve we should still clean it up.

Rather than setting that we need to destroy after we start the
component, we instead immediately set the destroy requirement on
creation.

Original-Fixed: 329497443

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007693
Original-Revision: 69031ee8a0f26e2f9549b6d593d38379dabf4d3f
GitOrigin-RevId: 11de99abe63cf7598698655bc321bcfe1b5836b6
Change-Id: Idd1d8817d934b118e1753892740e9f34a34dd47b