[roll] Roll fuchsia [starnix] Fix ownership of proc files
Previously, /proc/pid/* files were given the fs uid of the owning
task.

The proc man page says that they should be owned by the effective
uid of the owning task.

In practice, Linux seems to give some of the directories under
/proc/pid/ effective uid ownership, and the rest of the files under
/proc/pid/ real uid ownership.  In lieu of an explanation for that
behavior, we're going with the euid for the moment, as it is the
intent expressed in the man page, and closer to what Linux actually
does.

This CL also adds a (disabled) test that ensures that the files in
/proc/pid/ retain the same inode number across invocations. This
passes on Linux.

Multiply: verity-benchmarks: 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014460
Original-Revision: c5ad4a3a8895e8fdf1e4c7ac1e1609d60539b817
GitOrigin-RevId: 53bc263932ce9925dcaddcec48d044a9b81ca2f7
Change-Id: I1dd6d60e55527b8d980fa93267d2fbc3f02040fb