[roll] Roll fuchsia [Netcfg] Fix DHCPv4 Server "start_serving" bug
When Netcfg learns that a WlanAp interface has come online, it needs to
call `start_serving` on the interface's DHCPv4 server. This was done
correctly when the interface was observed to have gone from "offline" to
"online", but incorrectly when the interface is already online when
first discovered.

Original-Fixed: b/331036445
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1018296
Original-Revision: 65af04ca8e07660f98853e23c61931c104904382
GitOrigin-RevId: 31f147c08949be82636bd85db511ac38d27fa885
Change-Id: If2696343e6e8e13ba440f30345738d0892eaefca