[roll] Roll fuchsia [virtio-guest] Delete driver.virtio-gpu.disable kernel argument.
Previously emulator users set the driver.virtio-gpu.disable=1
to disable binding of the virtio-guest driver. However, this
feature is redundant as the users can always remove the virtio-gpu
device from the emulated device hardware configuration.

This change deletes the driver.virtio-gpu.disable argument so
that the driver will always bind as long as the virtio-gpu
device is available. This makes it easier to migrate the drivers
to driver framework v2 as we do not need to route the fuchsia.boot/
Argument protocol to parse kernel arguments anymore.

Test: "ffx emu start --engine qemu" on core.x64
Original-Fixed: 332552355
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1019132
Original-Revision: 65dad5586397365cf96821ca940abce65a127ca4
GitOrigin-RevId: 1949cc46190ff3eba788ab07b8fab73d3f41c9b1
Change-Id: I49e1215f50d64aa745e3145e92c565fd5cbe03c6