[roll] Roll fuchsia [bt][virtual] Use bt_log instead of zxlogf in Bluetooth virtual driver
This cleanup updates the logs in the Bluetooth virtual driver to be
bt_log.

This is part of the larger effort to migrate the virtual driver to DFv2,
which also decouples the LoopbackDevice and EmulatorDevice from the
driver framework.

Test: fx test bt-host-integration tests and checking test output logs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1018472
Original-Revision: 5eebda083e9e32df431e0225851b8b41ff9e42e9
GitOrigin-RevId: 2be18e7fa9bce90c3ed477ad3cdc8d98d479f535
Change-Id: I2530e7d2bb373cff4eab9e090bd7c666f500ccf7