[roll] Roll fuchsia [firmware] Fix `fx build sdk:firmware`
After https://fxrev.dev/1017115, the build was still failing due to this
collection with the SDK category "experimental" depending on SDK atoms
in the "excluded" category.

Move all SDK atoms, including the collection, that are specific to the
firmware SDK to the "excluded" category.
//third_party/android/platform/external/avb, which is in a separate
repository, is the remaining exception.

Also remove two of the atoms from the sdk_source_sets_allowlist since
they are not in the relevant SDK categories.

Original-Bug: 331832768, 331962496
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017794
Original-Revision: ea645081b47db96bae5b112e45335eb69c380345
GitOrigin-RevId: 755f33cb60b0582e5959be93187cef3d0ca2c84a
Change-Id: I4367fca52485fa10a501afbc0a31a21837c4c165