[roll] Roll fuchsia [kernel][x86] Move x86 PT code into header
This is a minimal cut+paste to move the X86PageTableBase methods to be
implemented in the header. Purpose of this change is to allow the class
to be templated without needing to move the implementation in the same
commit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013173
Original-Revision: 79b80e406b28e1b6b88548715cb38463379cbd3b
GitOrigin-RevId: 70e149f536b20467fd691b02005d439cd2ecd4f7
Change-Id: Ifae70f0e6bf6b93baa54340c59d90c8125fcc753