[roll] Roll fuchsia [sdk][vfs] Move Node class to `vfs` namespace
The `vfs::internal::Node` type is already a part of the public SDK VFS
API, and is used both in and out of tree. It is most prominent in the
`PseudoDir` type, which uses this base class for entry addition,
removal, and lookup. Therefore, it is moved into the top-level `vfs`
namespace.

Update enumeration constants to compliy with style guide.

Update all in-tree usages to avoid use of deprecated type aliases
and constants. Mark them as removed in the next API version to simplify
removal.

Test: No functional change.
Original-Bug: b/311176363

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987053
Original-Revision: af5427c7f601c5bc222b8a26a1f333955964c95a
GitOrigin-RevId: 306f20e39d18afe21b826cec6312729a06247381
Change-Id: I678f5776f508cc0f5c6a04385174860f2a42152e