[roll] Roll fuchsia [icu] Fix ICU flavor major version number expansion
Two issues.

(1) In the old code, when naming the ICU flavored resources,
the version number was interpolated early. This caused
*all* flavors of ICU to have proper resources, but
some would be placed in files with wrong names.

This ended with files from the "latest" ICU flavor, version
73 to be packaged in the file named icudt74l.dat, which should
instead have been icudt73l.dat.

This, in turn, caused the ICU loader to be unable to load the
ICU resources. This, in turn held everything up on the device,
causing the ICU tests to fail.

The new code names the resources correctly.

(2) The `icu_group` referred to ICU flavored deps using `public_deps`,
but the correct way to do so is via `icu_public_deps`. This ensures
that the dependencies are forwarded to the correct flavor.

After (1) and (2) were applied and I OTAd the device, the device
booted as expected. In the broken state, the device would stop at the
"G" screen.

Tested: OTA on a local device, confirmed that Opal starts.
Original-Bug: b/330351469
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1019994
Original-Revision: bfca95b72d719b5090ba1a437234241bee3742a9
GitOrigin-RevId: d140d1cb28ecc822d7a6d1e5974154073cc075cf
Change-Id: I0fdb15912dbf06832878a7730d512e542d7d2e84