[roll] Roll fuchsia [libc] Don't use a system call in abort
Remove the "fallback" code in abort that called zx_process_exit
in the event the trap instruction wasn't fatal.  In fact, the
compiler was already eliding the call because it "knows" that
it's "impossible" for `__builtin_trap()` to return.  It also
elides the loop.  So really there is no point in having anything
in abort other than the trap itself (and its ICF avoidance).
Removing it also makes it easy to eliminate the link-time
dependency on the vDSO (which per above was already not really
being used).  This was the source of the unexpected vDSO
dependency in the user.basic //sdk/lib/ld/test/modules modules.

Some of the test expectations had been set based on the extra
dependency being there, so fix those.  One of those was being set
that way inadvertently because the IN_PROCESS_TEST define had not
been plumbed to be visible in each test executable's own code,
only in the common test-start code.

Original-Bug: 42085436
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1019612
Original-Revision: acd4b64f7b555e9d9ef4d41ab5bd0e8927132853
GitOrigin-RevId: 6f04ca73416b7756e09adf6f548706d24fe7ae64
Change-Id: Id2ea0ccf8ba061e144cad3c100965a0f0122db44