[roll] Roll fuchsia [component_manager] Fix start ordering bug that contributed to starnix flake
There is a race in start ordering where StartAction performs the
following sequence:

- Create a pending Runtime object
- Call ComponentRunner/Start with this runtime object
- Install the Runtime object in the component's execution state

The race is what happens if the component exits between (2) and (3). The
component exiting will trigger a StopAction. If the StopAction runs
before (3), it will effectively no-op since it didn't find the
component's Runtime. But this isn't correct because the component _did_
start. At the very least, it should dispatch a Stopped event.

To fix this bug, change the ordering to:

- Lock the ExecutionState
- Call ComponentRunner/Start with this runtime object
- Install the Runtime object in the component's execution state
- Unlock the ExecutionState

Because the runner is called under lock, the race can't happen because
StopAction will be forced to wait on the lock, and when it gets the lock
it will find the Runtime there.

This doesn't completely solve the flake, but hopefully mitigates it.

Original-Bug: 325143035
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1018357
Original-Revision: 31770a65487491ef9bdcf7df2ae2d8a171f724f9
GitOrigin-RevId: d8f542c7044ad4c4929143c21e217001712414c4
Change-Id: I5dfcff4d3d8d078c0d39a73e00c58df419c04f42