[roll] Roll fuchsia [inspect] A Graph API for Inspect
This library intended for usage in Power broker and potentially in
component manager. It provides a Digraph abstraction with a common
schema. This provides the benefit that a single tool could be used to
convert this schema to, for example, dot and easily write tools and
visualizations around graphs recorded to Inspect.

The Digraph API follows Inspect RAII semantics.

Each vertex and edge of the graph can contain key-value metadata
properties representing state of the vertex/edge or metadata associated
with it. The values can be updated and overwritten.

The follow-up will introduce tracking of events representing changes to
the graph topology or the metadata associated with each node and edge.

Original-Bug: 331657401
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017412
Original-Revision: 04006ffd8c7c671b0b7cc66aa19333b813d78da7
GitOrigin-RevId: 784055818065d28da539b5515db4f1aef8fbef00
Change-Id: I77f37943884785ebd4a1cf84535d750f2741d9b1