[roll] Roll fuchsia [dl] Use ld::LoadModule base class
This CL plugs in ld::LoadModule as a baseclass to dl::LoadModule.

dl::Module is renamed to dl::ModuleHandle so as not to conflict with
ld::DecodedModule::Module, which refers to the AbiModule.

ld::SetModuleVaddrBounds and ld::DecodeModuleDynamic functions are used
directly.

Original-Bug: 324136435
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017118
Original-Revision: c0362bc767633e6244eb956203fae0110e032d75
GitOrigin-RevId: c8c05a17602a88790fc58cc3e3d649a4dc34f684
Change-Id: Ib61652d3120285b87f7f57c511e07a9a57a49ca9