[roll] Roll fuchsia [ld] Consolidate build ID handling
The ld::ObserveBuildIdNote function existed but wasn't being used
or tested; it was a buggy implementation that was needlessly
complex to reuse elfldltl::ObserveBuildIdNote for little benefit.
So rewrite it to be simpler, and use it in all of bootstrap.h,
StartupLoad, and RemoteDecodedModule.  This can later be reused
by libdl too.

Original-Bug: 42085436
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020512
Original-Revision: 3e7c29f5aff03f725199bb43d489c73d784ac35a
GitOrigin-RevId: 32b87486a67b842652fc1e17ab60781556ae21d0
Change-Id: I23f11aa926876eef19d7401e6f13a05f01a8d792