[roll] Roll fuchsia [fx][build] Print error for GN labels to unknown targets
Ensure that GN prints an error message when passed a GN label
to a target that is not part of the configured graph.

Before that, the label was passed as-is to Ninja, which treated
it as an absolute path, then complained it didn't know how to
build it. In other words:

BEFORE:

```
$ fx build //some:target
Building Ninja target(s): //some:target
ninja: Entering directory `/work/fx-better-build/out/minimal.x64'
ninja: error: unknown target '/some:target'
```

AFTER:

```
$ fx build //some:target
ERROR: Unknown GN label (not in the configured graph): //some:target
```

The pass-through was introduced in a previous CL [1] to deal with
GN aliases that appear in the Ninja manifest, but are not correctly
exported to the Ninja outputs database yet. However, this cannot
work for GN labels which begin with `//` and cannot be Ninja targets.

This CL fixes the behavior of the build/api/client commands
which take an `--allow-unknown` flag for this purpose.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1003432

Original-Bug: 42084668
Original-Bug: 293917801

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020792
Original-Revision: a91cf3358c7e10c3dc9091f61c89c9830ebb9432
GitOrigin-RevId: ac7f87d3569507d9fde9f5a9e5d55af559007e17
Change-Id: I48708ad075461f564ce2d9043145a87ea66b7869