[roll] Roll fuchsia [virtio_input] Loosen touchscreen condition
This effectively allows touchscreens other than the QEMU multitouch
touchscreen to be instantiated and bound to the virtio_input device
driver using our HidTouch touchscreen implementation. This is helpful
for using our virtio_input driver on crosvm emulators.

Prior to this change, virtio_input would read the device config to
deduce whether a device is specifically a QEMU multitouch touchscreen,
which the HidTouch class is implemented to support. This change expands
the condition to check whether a device has x and y dimensions at all,
indicating a touchscreen, allowing the driver to continue to bind to
devices of this type. Without this change, touchscreens that fail the
QEMU multitouch touchscreen condition will pass the mouse condition,
which is not desirable.

This change also adds logging containing device names and
the device types they matched with.

Original-Bug: 331970980
Test: manually tested in crosvm emulator
Test: manually tested on qemu emulator
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1019935
Original-Revision: 338a310751d541833011282a2df36d416405f9f2
GitOrigin-RevId: 6f78a0b0311deb177e5909fd6eaead605d6f683b
Change-Id: Ib43dc20c69ca94d3d3660de0169d4127c6ff2690