[roll] Roll fuchsia [netstack3] Support routine control flow
Support the Jump and Return actions in fuchsia.net.filter that allow
filtering routines to jump to uninstalled routines or return to the
calling routine.

This change also clarifies the difference between Accept and Return. All
routines evaluate to a final Return action by default, meaning that they
allow execution to continue in the calling routine (or to other routines
in the same hook). The default verdict for an entire hook is Accept.
When an Accept action is triggered within a routine, this action is
propagated all the way to the installed chain from which it was reached,
and execution continues at the next chain installed in the same hook (if
any).

Original-Fixed: 318718273
Test: netstack3-filter-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016954
Original-Revision: e0c990377ff58c666212ad2b8775c842384d5247
GitOrigin-RevId: 595be622ed1866172b402a35d9bdd1ae74bc4188
Change-Id: I356db4637cf756b831e4df6cec86463eb3830ddd