[roll] Roll fuchsia [http-client] Add a no-op idle timeout configuration
In the next CL we would be teaching http-client to stop itself if it
idles for some time. It should be well tested but to prevent any
unintended breakage we'll use a structured config entry to specify if
and how long the timeout is, and only flip the flag selectively.

This CL adds a config and verifies that it correctly falls back to -1.
This works thanks to the fact that adding a component to a package also
adds all of its dependencies, so the config value is propagated to any
tests that use this component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1019278
Original-Revision: cd2bd17d91ebec2bca181f7d5db278a8139c3daf
GitOrigin-RevId: bc812bf44b3c1d6f7f79a839e7dba8fa56fd2284
Change-Id: Ic91067af0b72036ebd42cb08c1f4925c35e87545