[roll] Roll fuchsia [netstack3] Dedup datagram socket info
Change the datagram `SocketInfo` to be essentially the same as the UDP
one, which contains concrete address types rather than associated types
defined on `DatagramSocketSpec`. The concrete address types are easier
to work with for the purposes of implementing getsockname in bindings,
and is also used for implementing socket inspect data.

Two side effects of this refactor:

1. ENOTCONN is returned when getpeername is called on ICMP echo
   sockets when the "remote port" is 0. This behavior was previously
   only true for UDP sockets, but since the justification is to copy
   Linux, the same thing happens for ICMP echo sockets on Linux
   (verified using sockscripter).
2. The port passed to connect on an ICMP echo socket was previously
   omitted from inspect data, but is now included as the logic for
   datagram socket inspect data is now generic.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013212
Original-Revision: dfa643602f8a88b76d3ff7cf277c5a5cecc04944
GitOrigin-RevId: f24da7119382de5e908d07118556dfce801ea943
Change-Id: Ib40c44829dd46f3fe1f29b953ad1b2adf535403b