[roll] Roll fuchsia [igc] Keep a counter for descriptor ring lengths
When the tx and rx descriptor rings become full their header and tail
indices will have the same value. This causes problems when
NetworkDeviceImpl::Stop is called because the driver can't use use the
comparison header == tail to determine when it has returned all buffers.
That comparison is immediately true without having returned any buffers
at all. An example of when this can happen is if a device is powered on
without an ethernet cable plugged in and then rebooted. The rx
descriptor ring will be full because there has never been anything to
receive. This caused reboots to be significantly delayed because
netdevice is stuck waiting for all buffers to be returned before
completing shutdown.

Fix this by adding a length variable to track the length of the
descriptor rings and use that to determine how many buffers to return on
the Stop call.

Also update the tx and rx ring tail index for each buffer in QueueTx and
QueueRxSpace to avoid firmware being confused by writing the tail value
to be equal to the head value when a single call to QueueTx or
QueueRxSpace fills up the entire ring.

Test: Manual tests to trigger these situations. Updated unit tests to
verify ring lengths.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021141
Original-Revision: e5204095fb2b769f88bdc673cfa91909652d8a60
GitOrigin-RevId: 5f47efa2dd65cbfa093e6b0e8d09f159c31ed757
Change-Id: If5db5cb27cc430cffa28706cc38a94fd8addaa3d