[roll] Roll fuchsia [starnix] Deflake starnix_sysrq_integration_test
This test launches a container with a crashing kernel and expects both
the container and the kernel to exit. However, it does so in a
particular order (container then moniker), and the order is actually not
deterministic. If component manager applies the stop action to these
components in the reverse order, the test will hang.

Instead, have the test wait for both monikers together.

Multiply: starnix_sysrq_integration_test

Original-Fixed: 325143035
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021125
Original-Revision: e9115a7e32d666848ad3362ba0cc43db76b6b325
GitOrigin-RevId: 0b8ee32cdfd3a14b560722e26f6bee9c9135e0c2
Change-Id: I2a1e9969faa1a9d86c2aa43bf1e9f09c72dab692