[roll] Roll fuchsia [dl] LoadModule::Load retrieves file
This CL refactors LoadModule::Load to retrieve the file. This is done
because retrieving the file is a common step for loading the main module
and dependencies; and it's practical to perform both in a single
function.

For now, the dyn phdr and needed count are stored in LoadModule to make
them easy to access during dependency parsing.

Original-Bug: 323419430
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1019337
Original-Revision: fd804a02b9ba3840bdf7a628eae534c8fff950ec
GitOrigin-RevId: 19c09537785e616e4972556f183e28b54feddb31
Change-Id: I8b2451e7607bdd696dc8edd14c6235519d11ccd4