[roll] Roll fuchsia [netstack2] Fix Flake in interfaces watcher tests
The interfaces watcher tests asserted that `Zx.ErrCanceled` is returned
when a watcher client is closed with a pending call to `Watch`. This
assertion is flaky because the call to `Watch` is run in a separate
goroutine, and it's impossible to delay the assertion until we know that
`Watch` has been called.

Instead, remove the assertion, as it's really just verifying a property
of the golang FIDL bindings, rather than the correctness of netstack2.

Original-Fixed: b/332987123
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021832
Original-Revision: 70df033fa08b4840cc88a56d3c878964d364b82d
GitOrigin-RevId: 5b889100399f716ae62759004dabd6f75fb82cc3
Change-Id: I868f48bde4abe3307363ab1407d31cbfa79deefd