[roll] Roll fuchsia [network] http-client can stop itself in tests
Adds a structured config arg to http-client such that we can configure
how long http-client waits when it is idle before exiting.

It uses the escrow APIs proposed in
I2f1ce4082d214b9c04e8dd6da62a700938697864 to park active connections in
component_manager, and arrange to be started again when either FIDL
connections or outgoing directory is readable.

This behavior is disabled in production and we'll activate it in
subsequent CLs, by configuring it in assembly.

The existing unit tests are multiplied over both the stopping and the
non-stopping versions and I also added a simple smoke test that sends
a bunch of requests and make sure they are all handled.

Original-Fixed: b/331997059
Run-All-Tests: True

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015133
Original-Revision: 5e42177457373272aa6a4c019def7ffd992be578
GitOrigin-RevId: 4ae8022cb5f2aaf997b4df0bb7b71e4c390cb44f
Change-Id: I9d80f6f0e86f775f9ea66fa9654a83edb83665e0