[roll] Roll fuchsia [debugger] Add more fine grained control of fd streamers
..to the Console. Now there is a similar counter as for input and
output, but applied for streamers. The Console::Stream method replaces
Console::WriteIfEmbedded, and is tied to the counter being enabled. Now
we are able to disable streaming where we need to without changing the
transition points for embedded mode.

The first use case of this is when we load symbols, so we can show a
helpful message so the user knows their test isn't stuck.

Original-Fixed: b/330771902
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013092
Original-Revision: 8beedd3494cea99699cb2bf95dc096ba926ea521
GitOrigin-RevId: d95c2b442d2e5a68a5038f531996a7647f0163dd
Change-Id: Id6613963ef9b1866ec065118db52848a6426b10f