[roll] Roll fuchsia [debugger][e2e] Pass target address verbatim
When dealing with ipv6 addresses, our logic was not doing the right
thing to encompass an address in square brackets. I think at some point
in the past this was important since we didn't always get the ipv6
address in the format that ffx wanted, but that's not the case anymore.

Original-Fixed: 333083347
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021856
Original-Revision: be9ce001a5b2f153f03287b36f16c605cbe0afd0
GitOrigin-RevId: fc6d04c1b69fe88ec144b2b5724164116a9b845c
Change-Id: Ib6408e8d14d754dd1305741472382e5ff0e299a9