[roll] Roll fuchsia [cpu_profiler] Acknowledge "Stop" Before Writing Data
If a caller is a synchronous fidl client, we need to delay writing data
to the output socket until we've unblocked the caller by Acknowledging
that we've stopped. Otherwise, we can deadlock with the caller waiting
for us to acknowledge and us waiting for the socket to not be full.

With the new kernel assisted sampling, we can now collect enough data in a test run to fill up the socket, running into the deadlock during tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020253
Original-Revision: 01b5d3c4b9d1020b89b1c283e1abf410894a125c
GitOrigin-RevId: 22455a580a667fad3303f9212110ccaf91a0b6df
Change-Id: Ifc86dd1aa56986314ff90adfe4496fbf8056ec1e