[roll] Roll fuchsia [inspect] Open validator protocol and remove unused methods
Go FIDL does not support open and flexible protocols. Therefore,
the VMO-based Go Inspect validator is moved behind a Rust intermediator.

The protocol between the intermediator and the internal puppet is
as minimal as possible. The intermediator now intercepts Unimplemented
actions on behalf of the Go component, as this doesn't have any real
bearing on what the tests are testing (the format).

The Validator test also queries the Archivist for the Inspect from
the component; the Go component does not export any Inspect. Instead,
the intermediator does that on its behalf after a call to
InspectPuppetRequest::Initialize. Again, this is because who serves the
VMO is not really relevant to what is being tested: the format of the
VMO.

The intermediator does nothing upon receipt of an
InspectPuppetRequest::_UnknownMethod. This means the protocol used by
the full-featured libraries (C++ and Rust) can continue being extended
without harming the coverage of the Go library.

Original-Fixed: b/330901977

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013033
Original-Revision: 3649a4d751e056516c6e64d0ae2b830f35436899
GitOrigin-RevId: f62ff94ce034ea882ab8f5687485dadc2983d910
Change-Id: I034bdacb4401373991162d1c208a8a380c0d057f