[roll] Roll fuchsia [starnix] Add tests for file creation.
This CL adds a few tests that call open with `O_CREAT|O_TRUNC`, expecting
to get the correct errors on certain filesystems.

If the filesystem is mounted as read-only, it should return `EROFS`.
If the filesystem is writable, but the user has no write access,
`EACCES` should be returned.
If the filesystem is writable, and the user has write access, the call
should succeed.

Original-Bug: 333037315
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021693
Original-Revision: c6791fc348023b834ab930fa54beecb4f3ede891
GitOrigin-RevId: 609cc22038ddf62d8c4f614cf1c9f706cf8ca851
Change-Id: I7bc84b3be0a640eec7697341c0bd1d867343662d