[roll] Roll fuchsia [kernel][lockdep] Improve circular dependency reporting.
Improve circular lock dependency reporting by displaying each direct
dependency of a lock that is in the same circular dependency set. This
makes it easier to determine how the circular dependency is formed and
which dependencies may be the easiest to remove.

Original-Bug: 333077530
Test: Run: 'k ut lock_dep_tests' and observe the correct dependencies
listed for the circular dependency tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021879
Original-Revision: dc703d6b1125664bdd956aeea9802e4a32216bba
GitOrigin-RevId: 0cd25167b02b4054f88d54714483fb5562b5d516
Change-Id: Ib5e6ff8499248fd56a02d3576b7f3a583ca59a4b