[roll] Roll fuchsia [kernel][sched] Have RunQueue hold onto the current thread
This change updates RunQueue to hold onto the "current" thread as state.
This is important as it will soon need to manage state relating to both
the running(/current) thread and the ready ones (e.g., for determining
the total 'flexible demand') - and it is easiest to do this when the
current thread is on hand.

This change also allows the current thread to be null, as there are
indeed cases where there is no eligible thread to run.

Two smaller consequences:
* EvaluateNextThread() no longer needs the current thread passed it;
* EvaluateNextThread() - having a name with tentative/'const'
  connotation - becomes a bit of misnomer with the introduced
  side-effect of updating the current thread: as a result, we rename it
  to `SelectNextThread()`.

Original-Bug: 328641440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1016493
Original-Revision: b84b6ea818604f837f4402eb8320d050fe4bbb14
GitOrigin-RevId: 15b8dd4bb7100f3334f55edd497328f25694ef86
Change-Id: Ia2a46cd9e402f1d7207955d49eb11e25907b222b