[roll] Roll fuchsia [kernel][sched] Update access to 'non-local' bandwidth parameters
This change removes the capacity, 'timeslice remaining', and expiration
accessors from ThreadBase, as the latter two are dependent on capacity,
and capacity is to become dependent on the state of the entire run
queue. Given that, it makes more sense to move access to this
information to the run queue.

Original-Bug: 328641440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1017461
Original-Revision: e08b89f7ce395da6e40dd33faf27a8c14290c0dd
GitOrigin-RevId: 10c567b2e73a7f2a5448592783f4873e929abfb2
Change-Id: I980848d48ec36eee762705118b4fbbc83d0c81f7