[roll] Roll fuchsia [storage] Add conformance tests for opening after rmdir
When we remove a directory, we should not be able to open it at the
path. `openat(fd, ".")`, where fd is references the removed directory,
is expected to pass. We should also be able to call fstat on fd after.

Fix bug in Fxfs where calling open2 with GET_REPRESENTATION attempts to
find the object linked to the removed directory returns early with an
integrity error.

Original-Bug: b/331491118
Test: fx test fxfs-platform-test
Test: fx test fs-directory-tests
Test: fx test fdio-unittest-package
Test: fx test starnix_user_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1015873
Original-Revision: e46459c3ce8fa9f1f6e42b6075d4bb065a3cd82a
GitOrigin-RevId: 63d171d05b6c0e9f4e39e2a748a78a27216aba32
Change-Id: I56a44a5cfd9a6d19ac944634f78ed7001b1e7b52