[roll] Roll fuchsia [build][python] Reorder union types tried during deserialization
When the serialization module is deserializing a value into a union
type, try to deserialize into non-string scalar and dataclass types
before trying to deserialize into the generic containers, and try the
string and None types last (if part of the union).

Example:

 some_field: Union[str, int, List[int]]
 value: [42]

This would previously have been deserialized into the string "[42]",
and not into a list.

Now it would be deserialized as a list of ints.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021396
Original-Revision: b2064b07b0593a4c4639db7e5df97f32f226bf13
GitOrigin-RevId: 05d97021313301a01941c7746b096187180c15e4
Change-Id: I359796cf8877876dfba306861efd37adac2d7b66