[roll] Roll fuchsia [idk] Reland "Build `vulkan_layers` only at LEGACY for now."
No changes, as this got caught up in a revert-stack.

`loadable_modules` (a.k.a., the "Vulkan layers") support being built at
individual API levels, but each of the subbuilds currently puts the
built libraries at the same path, so all but one of them gets clobbered.

My safer rewrite of "merge.py" noticed this, and is now throwing an
error. This change is a stop-gap mitigation to allow me to land the new
merger. Afterwards we'll take care of loadable_modules properly.

Having done a test build and diff, it appears that this doesn't change
anything about the final IDK itself.

Original-Bug: 316598209
Original-Bug: 327223536
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024012
Original-Revision: 9079e8f718765f4355d6c8b539822e713424b582
GitOrigin-RevId: c3d3868c951307e1b3bd32175fa35b68b272cf4e
Change-Id: I1f64c6775c3d688049d8be193ac934816bc9e911