[roll] Roll fuchsia [fuchsia-component] Support opening service instances directly
Existing methods allow clients to connect to default or named instances
of a service capability by providing a DirectoryProxy to a directory.
However, these methods expect the service instances to reside in the
service name subdirectory of the DirectoryProxy rather than the proxy
itself. This CL adds two methods to connect to a default or named
instance of a service located in a given DirectoryProxy.

Changes:
- Adds connect_to_service_in_dir() and
  connect_to_service_instance_in_dir() to allow connecting to instances
  of a service located directly in a given DirectoryProxy instead of the
  service subdirectory rooted at that proxy

Original-Bug: b/333433673
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1023401
Original-Revision: 1f9d0a5bd6198d23375743c4d0d6111c41e41db9
GitOrigin-RevId: 3b2a5f2db7f0523ec15722a299d40ee0bd6c9c2c
Change-Id: I7d0660454de4603a7fab6a359c3267fb7f3a09bc