[roll] Roll fuchsia [kernel][pmm] Remove disarm from checker
Removes the ability for the PmmChecker to be disarmed, and as a
consequence the ability to reconfigure it with different settings
during a single kernel execution. Doing this will allow for later CLs
to relax the locking requirements without losing correctness.

As the checker cannot be reconfigured, the existing host test is split
into two tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011452
Original-Revision: 13782a8001d627fcd5bc778f760cdc449a112687
GitOrigin-RevId: 9293cdfb0adca4d9497e53d5bba989ae67cbf2f9
Change-Id: I02da3bc0428ef8d3a0173fbe1d09275178770def