[roll] Roll fuchsia [kernel][pmm] Add additional scopes
Adds additional scope blocks to some of the guards in the PmmNode.
These scopes do not change any behavior in this commit, but will
make it easier for follow on commits to add code at the end of the
guard scope.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013192
Original-Revision: 997f6fe9931820c9514053be917caa42e5f51aec
GitOrigin-RevId: e089746e520a199e8f10d6a72c3019b71a93f677
Change-Id: I48ca4c276f955ed64b8854972ca6cc593f03ca6e