[roll] Roll fuchsia [netdevice] Use synchronized port shim dispatcher
It seems that using an unsynchronized dispatcher for the port and mac
servers in NetworkPortShim can cause a race during removal of the port.
The NetworkPortShim lifetime is tied to its server binding, when the
binding is unbound the NetworkPortShim object is destroyed.

The DevicePort class which owns the client end of the binding will call
NetworkPort::Removed on the client in its destructor, then immediately
after the client is destroyed and the binding unbound. There is some
evidence that indicates that doing this with an unsynchronized
dispatcher can lead to the unbind happening before or during the call to
NetworkPort::Removed. The Removed call would then attempt to use a
destructed NetworkPortShim object.

Change to a synchronized dispatcher for the port shim instead as there
is no reason to use an unsynchronized dispatcher here. Also remove the
storing of the binding refs in both NetworkPortShim and MacAddrShim.
These were never used and served no purpose.

Test: igc_tests netdevice-migration-tests gvnic-tests
      usb-cdc-ecm-unittest usb-cdc-ecm-integration-test
      virtio-netdevice-test network-device-lib-test
      network-device-microbenchmarks network-device-mac-lib-test
      network-tun-test wlan-drivers-components-tests
      virtio_net_unittests virtio_net_cpp_unittests network-tun-example
      netstack-integration-tests netdev-rust-client-tests
      netdev-cpp-client-tests network-device-test
Original-Bug: b/331057379
Multiply: netstack-socket-integration-test-no-err-logs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024114
Original-Revision: fc19ec06002720b871d723391d22e9e34f805a36
GitOrigin-RevId: 20f05ab79a28c50a51f43b76f4d75ba8b4e860e8
Change-Id: I407eaf52c0f54a437d9780116c577181c229251d