[roll] Roll fuchsia [inspect][contrib] Remove dependency on fuchsia_async executor
There was a dependency on the fuchsia_async Executor that caused this
library to not be usable in syncrhonous-only contexts. This change
removes this dependency.

As part of this change, the tests are now written using the FakeClock.
However, to use this fake clock we need to disable the parallel
execution of the test cases since there's one instance of the fake clock
shared across all tests.

Multiply: fuchsia-inspect-contrib-tests
Original-Fixed: 333603664

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024372
Original-Revision: fba44c4b7a85953363a580d3b8221990cbd35c32
GitOrigin-RevId: d4cdfa587af497bb2ebb1ef518ccb546aca37575
Change-Id: I74c19f0463648b115e0ea4b3b7777f79c92cd278