[roll] Roll fuchsia [netcfg] Remove empty variable declarations
The netcfg component and netemul testing component do not allow unused
results in the linter. Therefore, it is not necessary to specify a
return type of () from function calls.

This CL standardizes this across the netcfg files, including in
integration tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024892
Original-Revision: 6ef942b00053f3c823541d66ed1b095936587f03
GitOrigin-RevId: c984ef6207ecd15efecdb649e5773a694dda8622
Change-Id: I0ed6e317b37a9c1429e9ba8bbe773ab802f68a8c