[roll] Roll fuchsia [ffx][target] Improving target show schema
This changes the machine output for `ffx target show` to have a
more specific schema. Previously, the output was effectively a
map, now the keys are moved to be field names so changes in the contents
of the output will be reflected in the schema.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021395
Original-Revision: c1be77413fb75fd67f4c6b3dbc4fef6ef3e11a02
GitOrigin-RevId: 911f17a21fd1561da2ac5b5c2c44fd1b36c94ebb
Change-Id: Ic36a4903479fe628d411ed77d0f2cdec4d173839