[roll] Roll fuchsia [kernel][object_cache] Remove slab control size constant.
Remove the manually computed slab control block size constant. It is a
source of friction when changing the sizes of locks due to internal
structural changes or adding instrumentation.

Add a pair of static accessors that return the sizes of the control block
and slab entries for a given instantiation of the object cache. These
may be used by more advanced allocators to perform slab size selection
based on amortization thresholds or other constraints.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1023621
Original-Revision: 739d88b944cb1f808fd9fa70dc5ddeeb34621034
GitOrigin-RevId: 8ad09c9a188b950e47c9ec7ae6ffaab897b9a50e
Change-Id: I6b7e7b37d8642e233c330a6754d3a997d00a88ca