[roll] Roll fuchsia [ffx][process] Remove from being built-in
Removes ffx-process from being built in to the main ffx binary, allowing it to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes when `ffx process` is invoked, but makes it so that both the tool and ffx can build faster and with fewer resources making it easier to iterate on them.

Original-Bug: b/329326263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003813
Original-Revision: 813a2c539734935a71d09a8d47494309f4f049cc
GitOrigin-RevId: a09c7b8f20a44883c792d78e7d1b9387b51e5988
Change-Id: I59e35e39e6ad407bc0e5eba3f26d696e52a71daf