[roll] Roll fuchsia [time][rtc] Add an option `rtc_is_read_only`
This options allows us to configure a Timekeeper instance that
never writes to the RTC.

It may be helpful in some testing configurations such as the
one in the attached bug.

I also think props are in order to the original authors of this code,
whose design decisions way back when made this change very easy to do
today.

Original-Bug: b/333581050
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024131
Original-Revision: 2b018439f2b12552db5e6ee90c6655b60edf2b2d
GitOrigin-RevId: 6cb62306e3cf99c4b30bc17320df29e02a5b42bb
Change-Id: I4f3df96823dc545f07ba7d20559d59c5c7244cc4