[roll] Roll fuchsia [fidlcat] Pass string_view.data() to ifstream ctor
The ifstream ctor does not explicitly take a string_view as an argument
for its ctor. Prior to https://github.com/llvm/llvm-project/pull/85079,
the argument for ifstream was a `const std::filesystem::path &` so a
path is implicitly constructed. After that change, the ctor that
would've taken the string_view only participates in resolution if the
type is explicitly a `path` object.

This just instead passes the raw char ptr from the string_view which is
a valid ctor argument.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024987
Original-Revision: fc2ef428ec26b082ea18a4eda813ed5ba7dae74d
GitOrigin-RevId: 7f39bf7ec51301f0a448a73fb37ed1c0b8f1c8f0
Change-Id: Iff1caea4c0a3b0504cc9980f56bd3f871b333086