[roll] Roll fuchsia [paver] Simplify integration with paver lib
The previous abstraction adds a lot of complexity to the interface
between the paver binary and library. The library was more deply
integrated to help simplify things.

This change also paves the way (pun intended) for creating per board
variants of the paver, a long desired direction for the paver.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024975
Original-Revision: 4d48049385ae58a7bbe578162c510d6ddfdb4221
GitOrigin-RevId: 272c8ad2ea878dcb5561ba12e1655cf8fd78fc65
Change-Id: I52ca558255817e6b7317527507e062fce9b7c131