[roll] Roll fuchsia [ffx][trace] Remove from being built-in
Removes ffx-trace from being built in to the main ffx binary, allowing it to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes when `ffx trace` is invoked, but makes it so that both the tool and ffx can build faster and with fewer resources making it easier to iterate on them.

Original-Bug: b/329326263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003814
Original-Revision: 6a66dbea3eb2d9f1dc2592c30645572dc06658f3
GitOrigin-RevId: e6b056effb85bdff1948b0cb2c5eb2ef45094311
Change-Id: I415d47ecb0fdbae1d8663b764caaccab5c9cfd67