[roll] Roll fuchsia [kernel][arm64][periphmap] try to align mapppings to maximize large pages
Try to naievly align the mappings of the periphmap ranges to try to
utilize large pages. The algorithm is fairly simple so only really work
if the physical address is largely a power of two, but does a better job
than now, where it seems that everything will be misaligned in practice.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024927
Original-Revision: dd93089a4421e1c0675f1555da6f0a03e0896075
GitOrigin-RevId: 4b825933879433f7e29b07bb02180cb41a1832df
Change-Id: I43f49a955643b9adb9e5e122cf04b0e7ee0b2468