[roll] Roll fuchsia [kernel][vm] Fix old warning when reserving ranges in kernel
If a region is reserved in the kernel with anything but default cache
flags, deep inside the machinery that sets up a mapping of a fake VMO a
warning is printed that the newly created vmo's cache parameters
(CACHED) do not match the mapping (UNCACHED).

Add code to set the VMO's cache parameters before trying to map it.
Should have no functional change except get rid of the warning.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024928
Original-Revision: cbbcc0b79b79a2b4ef888927909fbb78c5ece816
GitOrigin-RevId: 400a08d4cf01a1510343ae3655b640485dabd32f
Change-Id: Ifb8edcc9de481fd99d5e8dbf89a20309ae400871