[roll] Roll fuchsia [netstack3] Fix bug in LocalTimerHeap
Caught this while migrating fragmentation unit tests to local timer
heap. If we have two timers at the top of the heap with the same time,
the heap would not reschedule the next wakeup because of the
optimization to not touch the timer.

Change the rescheduling strategy to be locally tuned to each of cancel,
schedule, and pop.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024982
Original-Revision: 59106f97aad3a129d5091c8b1e3679794bbdf020
GitOrigin-RevId: 512a4ae63ed3e7e37949d5afe383778f512c5f6e
Change-Id: I193ef85e4f2e5bacdd4386dc80a1ed73824bfbeb