[roll] Roll fuchsia [kernel][phys] Fix BootZbi::Error formatting
Make sure that the `.write_error` is not `nullopt` when
`.write_offset` is set, and same for `.read_error` with
`.read_offset.  This is necessary for the generic printing code
to actually show the offset in the error messages.

Original-Bug: 333770144
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025834
Original-Revision: cea28a45e2d0eba904e4684b87a0a2174fc026eb
GitOrigin-RevId: c86396ddff75c087d8db1e90dd9e55d01f9065a4
Change-Id: If7e266c9d9ca9f52c21735319ee2a987e76511f1