[roll] Roll fuchsia [fidl] Use infallible Endpoints<>::Create
This change migrates callers over to the infallible variant of
endpoint creation.  This reduces boilerplate
checks for the endpoint creation result.

This CL targets roughly half of the calls to
CreateEndpoints, in the fidl tests and benchmarks which
were able to be correctly converted by a script.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1026015
Original-Revision: 353d9eab4be61fe08bc2e4e556efefa833ab5c7f
GitOrigin-RevId: 4d07fb404f1547a0955d136e0a720d7149ab6372
Change-Id: I1ff2dc80a35d4e3395a6d4c3dd1eb76b3a20a9da