[roll] Roll fuchsia [display-coordinator] Clarify Client::SetLayerImage() logic.
This CL clarifies the use of an Image instance, vs the use of an
iterator in a HashTable that contains image. The CL also cleans up the
SetLayerImage error messages. This is a pre-factoring for future work
that reduces image_t usage in the display coordinator.

This CL does not introduce any behavior changes.

Original-Bug: 42078780
Test: fx test //src/graphics/display/drivers/coordinator/tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1026052
Original-Revision: 638d114a4b9758cabb5ac29115ffb74f64b4faaa
GitOrigin-RevId: 4eda873f57ff954f9fd7684493a539acb1df644a
Change-Id: Ib8aa3ebe404590fd66a74f26d80780b08bd2eb60