[roll] Roll fuchsia [media][fidl] Use infallible Endpoints<>::Create
This change migrates callers over to the infallible variant of
endpoint creation.  This reduces boilerplate
checks for the endpoint creation result.

This CL targets roughly half of the calls to
CreateEndpoints in src/media which were able to be
correctly converted by a script.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1026016
Original-Revision: 796ed8f655439a26b1115cc147345b9b104002fb
GitOrigin-RevId: dcdbab35d5f03e2e36f547262d3d72fb3ab20859
Change-Id: I86654e3454c41e3252e734887679884f1d027959