[roll] Roll fuchsia [kernel][phys] Fix exception-case backtrace formatting
This corrects the backtrace printing in the exception case so
that the frames are numbered correctly: frame #0 is actually the
interrupted state, not a collected backtrace frame; frame #1 is
the first collected backtrace frame.  It also ensures that the
header line before the backtrace is before frame #0 in the
exception case (when frame #0 is a special case rather than the
first of the normal cases), and further unifies this printing
code between the frame-pointer and shadow-call-stack cases.

Original-Bug: 333770144
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025974
Original-Revision: 2c0bd702bb1b7c129bcd55e16daa90a3dc69aab9
GitOrigin-RevId: 7a7264e2586ef3e9d91804c2c04dd92484b07aba
Change-Id: I1887746c7fe62c56b270190e460a63dc4bda0d5a