[roll] Roll fuchsia [ffx][log] Remove from being built-in
Removes ffx-log from being built in to the main ffx binary, allowing it
to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes
when `ffx log` is invoked, but makes it so that both the tool and ffx
can build faster and with fewer resources making it easier to iterate on
them.

Original-Bug: b/329326263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003812
Original-Revision: 7dc1ab18f73b6ed0e0adb3a9924bb918e41cb094
GitOrigin-RevId: 603b96022dbd82fc5b266fb20ebe96df5a43b50c
Change-Id: I93534999b0b02b7d0fbee88577042dca1939aabe