[roll] Roll fuchsia [drivers][audio] Reset signalprocessing completers
When a FIDL client reconnects to the aml-g12-audio-composite
driver after disconnecting, the previously pended completers
are not reset, which generates errors such as "WatchTopology
was re-called while the previous call was still pending".

This CL chooses `SignalProcessingConnect` as an appropriate
place (there are a few) to reset these completers.

Original-Bug: 321292607

Test: aml-g12-composite-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025904
Original-Revision: 9b7b2da94278102847fcbc06629e7260f13362e1
GitOrigin-RevId: 290ee32de34925a92a0dd4f2357fdf02fe1d035c
Change-Id: If3163cbf4f3887fad3178c476118aa6f47b462a5