[roll] Roll fuchsia [netstack] Don't add IPv4 limited broadcast route
We used to install a route to 255.255.255.255 on every interface as a
hack to allow the DHCP client to send broadcast packets as part of
initial configuration. This is unnecessary due to the client now using
packet sockets (which bypass the routing table entirely), and causes
broadcasts to get routed incorrectly (they should be routed using the
default route instead).

This makes this change on both netstack2 and netstack3.

Also adds an integration test that checks that broadcasts can be
sent / received correctly, which would fail without this change.

Original-Bug: 290794567
Multiply: ipv4-integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1020092
Original-Revision: 0584b329157b18580ef101d0aeb3a70d0f23abb5
GitOrigin-RevId: 46adf63286ef71a14fdd708df6f2de2faf5361f0
Change-Id: Ia75028a97d79d9e870efee678031ef980b6bb9ff