[roll] Roll fuchsia [debugdata] Support profile merging with version 9
There is an LLVM upstream change that changes the raw profile format:
https://github.com/llvm/llvm-project/commit/16e74fd48988ac95551d0f64e1b36f78a82a89a2

We have an issue in Rust on-target profile merging when we try to roll
Clang to a revision that uses the new profile format because debugdata
uses InstrProfData.inc that is shipped by the Clang toolchain. This patch adds support of merging profiles in the old layout that uses version 9 to unblock the Clang roll.

Original-Bug: 333945525
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025666
Original-Revision: a7017d4f26d517b868e65e80ad756d98337382a2
GitOrigin-RevId: 25f120c9f19c3262b23c6e6381e63747abf6256c
Change-Id: I9141f9d676a47b79c00354b1cf43998bc9ec28da