[roll] Roll fuchsia [component_manager] all ActionSet functions are async
In a future commit all of the functions in ActionSet will need to
communicate with a separate task, which necessitates the use of async
code. To reduce the number of lines changed in that future commit, this
commit makes these functions async and updates their callers as needed.

Original-Bug: 319095979
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1023952
Original-Revision: a206862f8b8baf0c66401ef2605ec0330dd3ad4a
GitOrigin-RevId: d935b235386c6e8c35e9fb3a24afbed8225252a1
Change-Id: I4a50bbd3f846def6abb2fcb297d3a42864b7d2df