[roll] Roll fuchsia [netstack3] Upgrade device ids w/o core context
Replace the functions in DeviceIdContext that upgrade/downgrade device
IDs with upgrade/downgrade functions in the WeakId/StrongId traits
respectively.

The old form was only there to allow fake device IDs to fail upgrade in
specific cases. Replace that feature with FakeReferencyDeviceId which
can be marked for removal just like the real device IDs used by the
device module, applying it to the handful of tests that check for it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027065
Original-Revision: a71861c5f14a67dba8510d798d7a60ff2b76d2f0
GitOrigin-RevId: e4be3e741e4cc72a7aacbce76c845c8ca0f9abb1
Change-Id: Ic529a96a483e29d215a6730f51d7bd5bfb1d4a19