[roll] Roll third_party/pigweed pw_transfer: Implement adaptive windowing in Python
This implements adaptive transfer windowing in the Python transfer
client. Window size is changed to be a multiple of the max chunk size,
and is increased in response to successful receipt of data and decreased
with packet loss. This follows an algorithm similar to TCP's
congestion control.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/147510
Original-Revision: e3c8e94cbffee472958ddd04c129494aed306ee3
GitOrigin-RevId: 404c0bbb9d2c9949ee33899785fa727dd091d680
Change-Id: I38a5fdafba868b1915029e3916ff7bec49e90e92