[roll] Roll fuchsia [build][testing] Remove ffx from data_deps of all tests
ffx was added as a data_dep of all tests, but currently isn't
actually needed. Remove it for now.

GN's data_deps are not ninja's data_deps, but are implemented
using order-only deps, which are always compiled before the
dependent target is compiled, when both are dirty in the
build graph.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028616
Original-Revision: 1ad672ce4343aa1ec157551fec5445493a8aac45
GitOrigin-RevId: b953438fa6e0923755093ba934c964d489e43745
Change-Id: I7e74cdf11a4727f282b32f838b0ca6b2bd865714