[roll] Roll fuchsia [component_manager] unresolve is no longer recursive
The unresolve action is no longer recursive. Before this commit the
unresolve action would recursively unresolve all children after shutting
them down. This work was unnecessary because these children would then
be immediately dropped by the parent when the parent was moved back to
an unresolved state, and this work was also undesirable because shutdown
is a terminal state whereas discovered is not.

Additionally, unresolving the shutdown children won't be possible in a
future change, as it won't be possible to move a shutdown component to
any state other than destroyed.

Original-Fixed: 334867885
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028293
Original-Revision: 052f26ff75bc0ece2807538da1f9d3b4e8a484ba
GitOrigin-RevId: 1e29e51bb81055fe4ae20c520c9301fec147229a
Change-Id: Ie2984fc8d05ce26b4c26d8b40823603a5102ce03