[roll] Roll fuchsia [ffx][emu] Remove from being built-in
Removes ffx-emu from being built in to the main ffx binary, allowing it
to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes
when `ffx emu` is invoked, but makes it so that both the tool and ffx
can build faster and with fewer resources making it easier to iterate on
them.

Original-Bug: b/329326263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003651
Original-Revision: c64257ef117698bd228b79a63922ebc4b714ebf5
GitOrigin-RevId: a6b7389e57db1d28c7b9b1b80f65679509f455c0
Change-Id: I96ef6e4769b788010177cd4977211d5d35431567