[roll] Roll fuchsia [tee] Translate results of TA calls back to FIDL
This adds logic to interpret the results of calling a Trusted
Application's OpenSession and InvokeCommand entry points and translate
them to FIDL types to return to the client. There is now a ParamAdapter
type that consumes a set of FIDL parameters and produces the TEE_Param
array and param_types value that the TA needs as well as remember the
types and details about the parameters in order to interpret the
results.

Original-Bug: fxbug.dev/332956721
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1023617
Original-Revision: f628744d0dc01da8e5f4daa0a8938fbf2cbad5f4
GitOrigin-RevId: 4f57b7fc38c22240be1cc230cd95f85081b6c8ec
Change-Id: I0dba966f311906809d0ca9474068c6fc6400e6fc