[roll] Roll fuchsia [package] Create a bootfs binary for exploring packages
This binary is like "ffx target-package" except that it is in bootfs
and so is available for debugging over serial. This will be useful
when the pkgfs shell directory that currently exposes all base
packages is removed.

To save space in bootfs, this tool and the "component" binary that
similarly mirrors "ffx component" are now implemented as libraries
that are linked into the same multi-call, busy-box style binary.
Copies of this binary named "package" and "component" are added to
bootfs, which deduplicates the files on disk.

Original-Fixed: 328455656
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027033
Original-Revision: 22ae75686e41724445c980cd682364ced4427afc
GitOrigin-RevId: d5c2a09613a7f6428af66099a771dcde5c15ff81
Change-Id: Iff0dffc70beaed503567ba49ad4b63bb8c1d532b