[roll] Roll fuchsia [botanist][syslog] Use `--symbolize off` with ffx log.
The --no-symbolize flag is being deprecated for --symbolize so we need
to update the flag (in orchestrate as well). This change also includes:
1. A fix for getting a new SSH client when the streaming fails.
2. CaptureSyslog() cancels the context of the syslog stream when it
   finishes so it doesn't keep looping.
3. We exit with an error if we get a certain number of errors within a
   certain time period because that'll likely indicate that there's an
   issue with the syslogger.
4. A tefmocheck to keep track of syslog failures.

This would've reported a flake in go/bbid/8751008761802843985 where the
`ffx log` command failed but it kept looping because there was no issue
with the connection to the target.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025573
Original-Revision: 12380341f9eafd4a43af34d4dcc69d8576eff8e3
GitOrigin-RevId: e1894cfc2ba154c7ffdc76fee178095446256192
Change-Id: Ie69982c75511d4b4529f004ac1abc4fa0b86797d