[roll] Roll fuchsia [kernel][sched] Introduce hybrid scheduling support
This change introduces the notion of flexible work: that which is done
when the firm work leaves bandwidth for it. Unlike in the kernel today
where a thread having flexible work to do is mutually exclusive with
having firm work, we introduce 'hybrid scheduling' in which a thread may
have both.

Original-Bug: 328641440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024515
Original-Revision: fbdc7bb83b9868d54e10612c677da5e56cf1aa11
GitOrigin-RevId: 0c8bbbe2e379ad75451983d511f877f698352755
Change-Id: I12dde9e8056cdf97f11b2fe6983ae4863d4a7c48