[roll] Roll fuchsia [netstack3] Call egress hook before device layer
The call into the hgook was previously in the device layer because all
packets have to traverse there, whereas there were a few paths out of
the IP layer. This change moves the call to the egress hook as well as
the entry to the device layer into a helper function.

This change also ensures, through a witness type, that it's not possible
to call into the device layer without having called the egress hook.

Lastly, because this change moves MLD and IGMP to use a new method on
the IpLayerHandler trait, CoreCtxWithIpDeviceConfiguration required a
few new trait impls.

Original-Bug: 328063820
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025673
Original-Revision: 44f60c41a031060eb099788531aa5f2fecf45f60
GitOrigin-RevId: 80d454d5b837e35bfd8c64d2380b411ba670d55c
Change-Id: I799a60e40b81e0feb54844109af9696a835abf1e