[roll] Roll fuchsia [zircon][sdk] Require explicit sdk category in zx_library
Previously, the zx_library() template allowed the `sdk_publishable`
parameter to be a Boolean or string containing an SDK category name. A
value of `true` was converted to "partner", and `false` was the same as
not being specified.

To simplify the logic and make the previously implicit "partner" SDK
category more visible, eliminate the Boolean options and require that
the parameter be set to an SDK category name if and only if it should be
"sdkable."

The `sdk_publishable` values added to individual targets do not change
the behavior and are not a judgement on whether that is an appropriate
category or even whether the target should be "sdkable." They simply
reflect the formerly implicit category.

Original-Bug: 333125197
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021875
Original-Revision: 727f482135a60a71a4ee0c8fd573f8ced418714b
GitOrigin-RevId: 24a1fb8eaa567d2dd9d1606ab64c4490197baab3
Change-Id: I539178b94bd6b555252c4c0ddc540693142a5306