[roll] Roll fuchsia [netstack3] Add IP-layer sidecar packet metadata
In the short term, this will allow us to associate a conntrack
connection with each packet.

The change was generated by adding a new parameter to the filtering
hooks and going where the compiler complained. It just so happens that
this manages to touch the majority of the IP layer and a small part of
the device layer as well.

In the long-term, we'd like to be able to pass information across layers
(e.g. from transport to device).

Original-Bug: 328063820
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1021352
Original-Revision: abd7c3968968851b41a02763f5b26772028ea82d
GitOrigin-RevId: 69a2de559ea6fadb04050bb62fc88db4a5244263
Change-Id: Iccb259063952bb7b7d53da08baf6fa9cd393fd6b