[roll] Roll fuchsia [Honeydew]: Wlan typing and wlan affordances fix
This change fixes KeyErrors in the wlan and wlan_policy affordances by
switching over to using get(). It also fixes two typing issues. One with
the ProtectionDef returned by the wlan facade which is converted to a
string by serde in Rust. The other fix is changing ClientStatusResponse
to a protocol to better represent the fact that only one type of
ClientStatusResponse exists at a time. Finally, it updates and fixes
some of the unit tests for wlan.

Verified the following on Patchset: 5
* HoneyDew code guidelines
* Successfully ran the impacted functional tests using InfraRun

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1024592
Original-Revision: 5e78798409a87f8bc1d64afeb92e0d34e563789c
GitOrigin-RevId: 73ffa78838ee2a66370ccb1ab12205e6af8f8571
Change-Id: Ief385352ca2f9573d5f39a055d8d6f24c845f92a