[roll] Roll fuchsia [kernel][vm] Do not Resume a thread while holding spinlocks.
Do no call resume on a thread while holding spinlocks, unless local
preemption has been disabled. If we are holding spinlocks when we call
resume, and the thread we are resuming is assigned to our current CPU,
it will trigger a preemption event.  It is illegal to allow a thread
to become preempted while holding a spinlock (otherwise we might end
up with non-running threads holding spinlocks, which would be Very
Bad).

This instance was missed, probably because the code in question only
runs in builds with the debug compressor explicitly enabled (which,
AIUI, is not many builds).  Even when this happens, the odds of
hitting the failure are relatively small as they require the system to
be sufficiently loaded to the point that the best CPU to put the
resumed thread on is the current CPU.

So, add an ASSERT to Thread::Resume meant to catch any attempts to do
this in the future (regardless of whether the operation would have
resulted in a local or a remote reschedule), and add an APD around the
spinlock scope to insure that any local reschedule operations are
postponed until after the spinlock has been dropped.

Original-Fixed: 335213526
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029435
Original-Revision: 0ae56f6633eb8a9374e8da5c29508160dc14cc88
GitOrigin-RevId: fb0f5d0bc78e509c253f5651d4c28690ed29564c
Change-Id: Ia44873a65fcd91d5cde14e2065b50f03bfdf1566