[roll] Roll fuchsia [kernel] Improve immediate access marking
Splits the MarkAccessed method into having an early before the lock
short circuit that is inlineable, and then the longer locked remainder.
This short circuit obviates the need for the reclamation checks in
UpdateOnAccessLocked, simplifying the logic.

The pf_flags passed in to UpdateOnAccessLocked had become redundant, as
LookupCursor is already skipping the call in the case of hardware page
faults.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1026252
Original-Revision: 41bced5e9482a31b967d87be7c11f91ba5f62166
GitOrigin-RevId: 10f0ebe181de91f69177ef5ccd90d9e7502bea3b
Change-Id: I3408a170a373834ae15c3b8184c3b3c37f615e29