[roll] Roll fuchsia [idk] Make idk rule hermetic.
This involved a slightly annoying amount of refactoring of the merge
process so that all the possible input files were made available to the
toplevel script, but I actually like it better this way.

Confirmed that it's a null-diff in the generated output (except for
helpdoc_out.tar.gz, which always seems to change).

Original-Fixed: 333054435

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028373
Original-Revision: f3f16dac2a47309ba7e96007b818cda1ff21a078
GitOrigin-RevId: 7ea87fc5a0d97a5f7cd171c888c77ae62a246eb9
Change-Id: Ica71f4bb6bde678524f4fb4216912735b58c0dcf