[roll] Roll fuchsia [component_manager] resolve waits on discover, refactor lifecycle test
The test model::tests::lifecycle::bind_action_sequence was asserting
that a parent's resolve event must arrive before a child's discover
event, when there was no such guarantee in the events system.

The child's discovered action is started as part of the parent's resolve
action, and before this comment which action triggers an event first was
not deterministic.

This commit causes a parent's resolve action to block on a child's
discover action, which guarantees that the child's discover event is
emitted before the parent's resolve event.

Original-Fixed: 335118565
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029272
Original-Revision: de6f46c962aff0cb6bf9f6730d09f3989cc3925b
GitOrigin-RevId: b9da84699a5c6ac8753d62b4e09d3c241dded23b
Change-Id: I8475e884b8a4f3ff928fe0de7939d5c7e27727c6