[roll] Roll fuchsia [netdevice] Improve test RX queue event handling
Various stability improvements to the network device lib tests.

The WaitStart method's RX queue trigger handler could sometimes get
triggered by a delayed kSessionSwitchKey packet. This caused WaitStart
to return too early. Solve this by ignoring any kSessionSwitchKey in the
Start handler and only signal completion on the very first kTriggerRxKey
packet. Some tests also did not correctly call WaitStartwhen they should
have. This could cause the RX queue trigger handler installed by the
test to be overwritten by Start's event handler. Make those tests wait
correctly.

Several tests wait for a kTriggerRxKey packet as an indicator that
CompleteRx has completed. These tests would sometimes fail because they
would instead observe a kFifoWatchKey packet instead. This is part of
the normal flow as Session::CommitRx, which is part of the call to
CompleteRx, will write to the FIFO which triggers that packet to be
sent. Fix this by simply ignoring FIFO watch packets. In order to make
this easier for all tests involved this change introduces a shared
helper method.

Fix an issue where the RX queue trigger event handler could sometimes
get called after a test had gone out of scope. This happens because the
event handler is kept alive for the duration of the call to the event
handler. So even if the event handler was cleared at the end of the test
an in-progress call to the event handler could complete afterwards. This
would then incorrectly access data on the stack of the test. Fix this by
having the helper method allocate a completion on the heap and capturing
it in the event handler. The helper method then provides a raw pointer
to the completion that the test can use.

The event hooks used for testing were never thread safe. This seems to
have mostly worked but to be on the safe side this change adds locking
so that event hooks are always modified while under locks. In order to
preserve performance the presence of an event hook is checked with an
atomic load with a relaxed memory order. On some CPU architectures this
is as fast as the previous implementation and on others it should
hopefully be close enough that it doesn't matter.

Original-Bug: 331044823
Multiply: network-device-lib-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029895
Original-Revision: 4374e5f20d522cf55139204c03ba6fe82a8b8127
GitOrigin-RevId: 9c043a233faea3a532d6b11dd8c92c4fb4adda38
Change-Id: I9b2380571f30a68d1dc523617c66e79ccb3a900a