[roll] Roll fuchsia [rfc] Explicit External / Platform Split in SDK Interfaces
The Fuchsia IDK doesn't specify whether products should implement clients or
servers for the various FIDL protocols that it defines. This makes it difficult
to correctly evaluate the backwards compatibility implications of changes to
Fuchsia APIs. This proposes a way to express that information in a both human
and machine readable way in FIDL. This unlocks opportunities to improve the
correctness and stability of Fuchsia-based products.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982353
Original-Revision: 320ffb4d147591e896b7bd10065636c910453b24
GitOrigin-RevId: 4f5864ce8c8446292c205d2130b6d3fdc5c8e762
Change-Id: I079a4907fe223243d6ff5257e6d993c41b178f0d