[roll] Roll fuchsia [ld] Introduce MockLoaderServiceForTest
This CL moves the logic of controlling the MockLoaderService from
LdLoadZirconLdsvcTestsBase to the new class MockLoaderServiceForTest.

The LdLoadZirconLdsvcTestsBase API becomes thin wrappers around the
MockLoaderServiceForTest API and is entirely defined in the *.h file to
hold the trivial code.

MockLoaderServiceForTest implements the business logic that was formerly
in LdLoadZirconLDsvcTestsBase, accepting a `GetVmo` template parameter
to specify the callback function that will return the VMO result for a
given `name`. It will lazily initialize a MockLoaderService. This API
can be shared with libdl and libld.

Original-Bug: 324136435
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028354
Original-Revision: a606e835d38ea5e1bf79cb458df96afb2f4329ca
GitOrigin-RevId: d27cc595c359879ce397403a4e974a04bf4d8a9c
Change-Id: I7386e1cb7f1f1000d807b1e3c6dde81bc17d0e65