[roll] Roll fuchsia [goldfish][pipe_io] Use zxlogf() macro defined in the Driver Framework migration utils library.
The goldfish pipe_io library provides helpers classes for goldfish-*
drivers to read and write contents to the pipe. Currently, the library
is depended on by the goldfish-sensor and goldfish-display drivers.

The library uses the DDK zxlogf() macro to print logs, which is
not compatible with the Driver Framework v2.

To make it easier for DFv2 migration of goldfish drivers, this change
replaces the DDK-defined zxlogf() macro with the one defined by the
driver-framework-migration-utils library, which can be linked against
a DFv1 or DFv2 implementation.

Test: no-op expected.
Test: virtcon and vkcube-on-fb work on FEMU (ffx emu start) on core.x64.
Original-Bug: 335325419
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030192
Original-Revision: e9201abcb369f16d7db93fc76e6c63bb4eecb179
GitOrigin-RevId: 8fb211bb4b42cfd6ed98844f6a9af46afde5e433
Change-Id: Ibb9bc65518d6210fefcdfd297ca668f3f05299f1