[roll] Roll fuchsia [kernel][vm] Explicit accessed interval
Adds kernel cmdline option for explicitly controlling the accessed scan
interval instead of forcing it to be related to the min aging interval.
This is intended to allow for lowering the min aging interval without
increasing the mandatory accessed scan rate.

There is presently no control in the system assembly for changing the
min-aging-interval, so by defining the accessed-scan-interval default
value to be what the code would previously calculate there is not
behavioral change to any products.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1025552
Original-Revision: 2b84ea2e67028bf6ef5649f5e158f353fbaf10dd
GitOrigin-RevId: 959b89e951899e7f4e8ec56aec56b49fc80ebd58
Change-Id: Ie77e3d1ffef3922edaeb40097bd695db787e035a