[roll] Roll fuchsia [mock-mmio-reg] Fold the -gtest and -zxtest rules together
This adds a new :mock-mmio-reg-zxtest target to the mock-mmio-reg
library. This target is functionally equivalent to the lib found at
//src/devices/testing/mock-mmio-reg-zxtest. Rather than copy the code,
we can just tuck the differences (which really boils down to whether
zxtest or gtest is used for the ASSERT_X macros) behind a -DFLAG build
define and #ifdef block.

The plan is to move all users back to this (new) target and only have
one version of the source to maintain. Once everyone is migrated back to
this target, the code at //s/d/t/mock-mmio-reg-zxtest can be deleted.

Original-Bug: b/328812587
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029799
Original-Revision: 115b5ba821e15599f0de229f084015d1d549ab77
GitOrigin-RevId: 42f2d244eee2732cddf56d310acc5a3929552ad4
Change-Id: Ie203097aa4857536daab1f9f85c1a66febf9f96a