[roll] Roll fuchsia [netstack3] Avoid dynamic dispatch and alloc
This has been bothering me for a while, it's rather trivial to implement
this iterator for device addresses and it saves us from a box and
dynamic dispatch.

This is still a problem in Slaac, but this is one unneeded dynamic
dispatch down.

Original-Bug: 42059236
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031157
Original-Revision: b9f98c24e0dbe1552166931999466168de2fee6c
GitOrigin-RevId: 0d6da88f9455784221245bc6bc4ecdf1dba67470
Change-Id: I05a7dc13e62d64f8174fc2844fdc60cf7392ebe6