[roll] Roll fuchsia [netstack3] Include routine/rule counts in inspect
Currently, when a hook has no routines or a routine has no rules, the
inspect data published by Netstack3 shows it as an empty node. This
makes adding an inspect selector for that node difficult without also
blanket selecting all properties contained by the node. Add properties
to the inspect data that indicate the number of routines installed on a
given hook and the number of rules in a given routine, so that there is
a property we can select on in privacy filters for an otherwise empty
inspect node.

Original-Bug: 332702047
Test: netstack3-inspect-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031314
Original-Revision: d27a773fad3ec21a365192b614ee991df20eb2f1
GitOrigin-RevId: d2f2b2b297722796cb8e5d00fe62b7b895704f6d
Change-Id: I367dd96ceac5c4102ac29ca48a91dce80b8968cb