[roll] Roll fuchsia [netstack3/conntrack] Basic generic state tracking
This implements state tracking that isn't specific to any particular L4
protocol.  For now, this means connection state (unestablished or
established) and the time the last packet was seen on the connection.
This will be especially useful for the table size limit
(https://fxbug.dev/328064032) and connection GC
(https://fxbug.dev/328063959).

Because this requires getting the current time, there's a lot of
plumbing to get FilterBindingsTypes and the new FilterBindingsContext
into the conntrack module.

Original-Fixed: 328064022
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031172
Original-Revision: 9dbef4444b1103d841f96f226ad2542d06a639fc
GitOrigin-RevId: 31078ea222897f40eeea7047662956bdd0bd32eb
Change-Id: Ie84ab3762775f8ad3d67a4eb23f1c9875ee219fd