[roll] Roll fuchsia [starnix][selinux] Create a clearer layout for the LSM/SELinux impl
The kernel's "selinux" subdirectory is renamed "security" to reflect
that the hooks and data structures it provides are in principle generic,
not tied to a specific choice of Linux Security Module.

Within the "security" directory a subdirectory holds the SELinux-specific
integration for the various hooks; since SELinux is the only LSM built
into Starnix some aspects are currently (re)published from the "security"
module.

The SELinux and procattr filesystems remain in the "security" directory
for now; the latter will be migrated to interface to SELinux via hooks,
similar to other parts of the kernel, and moved into the kernel's "fs"
module.

Original-Bug: 335397745
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030212
Original-Revision: 736cd3bc644950d7dc60bdd5dbb978c3be845a41
GitOrigin-RevId: 91656d862cad93e00af19e3a31e36b4a68330a13
Change-Id: Ie15fc56c7f0f1413f3a20f88f127c77542705b8b