[roll] Roll fuchsia [wlan][policy] Removes WLAN power control.
The low power listener is not used in practice and the
PHY APIs that the current implementation relies on are
going away.  This change removes the linkage to the power
client level service and the ability for wlancfg to set
power mode on PHYs.

Test: Flash DUT and ensure that it boots.
Original-Bug: b/332904477
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030824
Original-Revision: 95d1a589a416ddafc11b6a45b9b263c84afd2fdd
GitOrigin-RevId: b9131188444cb3f9361853b55514e9a73946a511
Change-Id: I84c26d0118ed9d6161d22a7be8bcd3649321fd6f