[roll] Roll fuchsia [bazel][rbe] Add missing profile runtime lib
This fixes ld.lld link errors in coverage/profiling configurations.
The profile rt-lib variant is currently added unconditionally, even when it
is not needed, but there is not much harm in this.

There is only libclang_rt.profile.a, there is no .so.

Original-Fixed: 335704815
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031235
Original-Revision: 50cd2e72bf5ce391374e769e85519c5bd1b8c831
GitOrigin-RevId: 73c9f5628f4e798dd6bafe598b03c03a0c33b0ab
Change-Id: I618daaf8e6c310c43bc4a65972f939345b799ce7