[roll] Roll fuchsia [assembly] Deserialize json5 using serde_json temporarily.
serde_json5 allows a syntax that serde_json doesn't, so this change
has assembly temporarily bounce the deserialization through a
serde_json::Value, and then serializes/deserializes that using
serde_json so that there is a backstop against uses of the improper
enum syntax that serde_json5 allows.

This is to be removed when serde_json5 is corrected to match
serde_json.

Run-All-Tests: true

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031282
Original-Revision: bb11715b29d7ca0181557ed9efcd860edaafb385
GitOrigin-RevId: efd880244da51d8309d367043a2c9290529fb519
Change-Id: I74f16ad70847ca0d1ea32111d718a6279fdc5693