[roll] Roll fuchsia [build][tests] Don't generate errors when developer_test_labels contains shell commands or config data
Move the dependency edges such that the developer_test_labels is
not subject to the assert_no_deps checks that discoverable_packages
are.

This addresses complaints about various test groups not being
able to be used with the 'fx set .. --with-test' option.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1032037
Original-Revision: 827d9ca216d0a143b6354602f2ab65337cf9281c
GitOrigin-RevId: 6bd5e5a23c8473fbfed1fca79d60af77c1db8084
Change-Id: Ib0d3bdb2c92f64b6ddb3a11a39d5c7a6259bb3a8