[roll] Roll fuchsia [audio] Log device removals/errors
Log when a device has an error, and when it is removed,
in a consistent way to how we log when it is successfully
detected/queried/initialized/added.

Also set the default log settings as desired for integration
of a new non-DevFS Codec driver from the Bluetooth team.

Test: audio_services_device_registry_unittests
Test: audio_services_device_registry_warning_unittests

Original-Bug: 321292607

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030822
Original-Revision: 7cbfd52f25157410c68cdf9ad559afc93deb128b
GitOrigin-RevId: bd00d0eefc3e1b3a414dccc0714cfedce92a6205
Change-Id: I58af6e1fc47c4c8fd91fdfd976b66a411a42c3fa