[roll] Roll fuchsia [build][assembly] Move cmdutils group into assembly
This moves the 'time' and 'uname' shell commands' packages into
assembly, and drops cowsay and hwstress from the group, as they
are not used by tests, but their tests are included.

The group itself will be removed in a later cl.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031917
Original-Revision: 29fcc49f5baaae89b8338cde282237a12550d56d
GitOrigin-RevId: 32b3613f85b1b7094c948dc4e61d831db2d2550b
Change-Id: I739e58064943bc47fce8851c9e11674ba8426903