[roll] Roll fuchsia [aml-pwm] Allow arbitrary number of pwm id metadata
The existing code expects a specific sequence of pwm id to be passed in
as metadata. Updated the code to move the assumption on pwm id into the
driver and only validate the incoming pwm id metadata on whether it is
supported on the current board.

Test: aml-pwm-test, device enumeration test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030498
Original-Revision: f8e6e9fe6ee3d502f56eae2e3eb8c763c1ac3bd3
GitOrigin-RevId: 925d0d02da5decebdd0cbc10371325fc399d3384
Change-Id: I1ec277550e0a1b177b24ef01a763d58f94c66749