[roll] Roll fuchsia [scrutiny] Add minimum_utc_stamp presence check
This check ensures the existence of the minimum_utc_stamp file as a
bootfs file at config/build_info/minimum_utc_stamp. This file contains a
time stamp corresponding to the latest commit from which the build was
created. The file is depended upon as a measure of build age.

Original-Bug: b/333092690
Test: fx build (with scrutiny-enabled target)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029873
Original-Revision: 2c4bd0034533080630010fea2cbdafab058a29e2
GitOrigin-RevId: 52775bb81e0b769b64afa25335f6d975b2f2c539
Change-Id: I2c1142940563691a759c774994d5066d66a2f777