[roll] Roll fuchsia [bt][examples] Add basic tests to the LE battery service
Add some tests to verify the common paths for the LE battery service.
This also means that LSC and API changes will be caught in CI/CQ for
this component since it is not included in any configurations.

Original-Bug: b/326423915
Test: fx test bt-le-battery-service-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1033923
Original-Revision: be8c8442da8bf080ce55c59512c236a75896e564
GitOrigin-RevId: a9e45c4e4e1ba751710009dd2614fed8257bb41a
Change-Id: I3b977eb7de87f921084bb91489ce01fb7f83bd41