[roll] Roll infra/recipes [build] Notify when builder did not rebase change if failures occur.
Since integration is checked out at HEAD but only the patch is not
rebased when `do_not_rebase_patch` is true, build/test failures can
occur if the patch is out-of-date with the rest of integration. This may
not be a problem with the CL itself and just needs the CL to be rebased
on top of HEAD so mention that in the failure summary.

Original-Bug: 335727014
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1032020
Original-Revision: 60e4c0daf17419e102636f2cbd8807e2e2a9860e
GitOrigin-RevId: 67fbb5031c0b84163f6a8bc46a22bfdfbe801f9f
Change-Id: Iddae0220a24fef0a2da4dc472ffb5ad75a04968b