[roll] Roll fuchsia [fake-display] Refactor layer and config lists in unit tests.
This CL is a prefactoring for I0bf6dca1f806a0c99bdd347779f7eb4057a927d9,
which migrates the Banjo API between the Display Coordinator and display
engine drivers away from lists of pointers.

This CL migrates the lists of layers and display configurations
hard-coded in tests away from lists of pointers to disparate data
structures, and towards arrays of structures. The lists of pointers
required by the current Banjo API are built on top of the arrays, and
can be removed in an easy-to-review diff when the Banjo API is migrated.

This CL also switches the fake-display test target to
fake-display-unittests, to match the convention that is emerging across
the display stack.

Original-Bug: 314126494
Test: fx test fake-display-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034116
Original-Revision: 9c37cc194ebbcaf2bcaef91358743cd03931c970
GitOrigin-RevId: 87d8784c4f7b2497eff5062aeca01e06ba18ef5a
Change-Id: If3a2e14cb1cfa92129c4bd23afabc6f454adfa2a