[roll] Roll fuchsia [assembly] Introduce a new feature set level: Embeddable
This new feature set level is aimed at use cases where we don't expect
fuchsia to be self-updateable, but rather, externally updated. This
affords us the ability to drastically shrink the set of dependencies
necessary.

It is also meant to be the basis for products which are as small as
possible. It strips out the core realm entirely and will only bring in a
subset of the bootstrap realm. Features like power management, thermal
control, diagnostics, etc, are all considered optional add-ons.

Embeddable is currently a complete subset of the bootstrap feature set
level at present, although that may change in the future.

Original-Bug: 325104472
NO_IFTTT=Lint is incorrect
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027172
Original-Revision: 9c7938b66ac5bb21d968543e80453df06405afe5
GitOrigin-RevId: 9b7ca6b9d05810752f974a27f107bfe0a88c3388
Change-Id: I41f4b18edfcdee14b7d3a8e4a0377961136753c6