[roll] Roll fuchsia [rfc] Raise user-defined Zircon exception
This RFC introduces the `zx_thread_raise_exception` syscall, which
raises a user-defined Zircon exception. The first use case for this
syscall is for Starnix to signal to the debugger when one of its
processes calls `exec()`.  The debugger will use this signal to
determine whether the developer wants to attach to the process.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1013525
Original-Revision: a6b945a01cb2a09f0703745dadd49293597f7f4c
GitOrigin-RevId: 0480412bd792466dd2a1160d348408e88cbb114c
Change-Id: Ia256e6f56ad92e2cf9978ca3432188fcbc164e58