[roll] Roll fuchsia [wlan][policy] reformat new file for storage
The new file storage for saved networks is implemented but not
used yet. This CL changes the format of the new storage, and the new
storage is still unused after this change.

The previous format was a map of network identifiers to a list of
config data with that network identifier. That format was made to match
the previous way things were saved in stash and the way things are saved
in memory. For example, stash would write a value to a key, so the map
format made sense, but with the new file version the entire file is
rewritten each time.

Test: updated unit tests for the reformatted storage.
      (fx test wlan-stash-tests)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1014465
Original-Revision: 95762481c03e8822fe9f8ae2df51be74487487fc
GitOrigin-RevId: 30bd08fc56627dd3bf9bd6abbcf7e8bca303aa02
Change-Id: Icf379d189107d977266808bde7453230788f06a7