[roll] Roll fuchsia [fuchsia.io] Downscope connection rights based on protocol
Update protocol to document that the rights granted to a connection
can be downscoped to the operations which the target node can possibly
support. These rights can still be requested if the parent connection
has them to support protocol resolution.

This makes the resulting behavior more consistent with the current
implementation. Previously, opening a file with OpenFlags.RIGHT_WRITABLE
would only grant Rights.WRITE_BYTES + Rights.UPDATE_ATTRIBUTES to the
connection. However it was possible to obtain Rights.MODIFY_DIRECTORY by
specifying that right in an Open2 request (or by specifying no rights).

Report which rights a directory supports opening child entries with as
part of the node's abilities, and document that these rights apply to
both files as well as children of the current connection. For now we
update the conformance tests to assume directories support all rights,
but in the future this may have to be moved to the test harness (e.g.
the C++ VFS does not support client-side mutable directories).

Add TODOs to track disallowing specifying no rights with Open2.

Original-Bug: b/324932108
Test: fx test rustvfs-io1-conformance-test fxfs-conformance-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031188
Original-Revision: 7a2adbd7836a7b0c09c692c25e5e92446366d755
GitOrigin-RevId: b70e18bd18fad222642bd2a15a33e84e09828841
Change-Id: Ic0f1db570c06001314fbd6be1909f14212bd64fd