[roll] Roll fuchsia [fxfs] Makes profile replays and recordings happen together.
Previously either a recording or a replay would begin exclusively
but now a recording is always begun and if a replay is available then
that is also replayed at the same time. The recording will start in the
graveyard and only be placed in the profile directory when properly
written and completed.

Original-Bug: b/333552162

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031182
Original-Revision: 088b1323cc494d6a1a3913c09e9f484049c616f6
GitOrigin-RevId: 0dfd2b0fc150471bcf517c0d787779a64537c317
Change-Id: I623e4733e07ddfd3b00b1b86471915a4519446e3