[roll] Roll fuchsia [starnix] Treat kernel processes as critical to the job.
Only terminate the kernel's job if a process exits with a non-zero code
so that normal termination of a process won't kill the job. The only
time that Starnix should have a Zircon process exit with a non-zero
code is when there's an unhandled exception that bubbles out of the
kernel's job. Regular process termination and Linux signals that
kill a process should not appear as non-zero Zircon exits.

This replaces the use of the kill_job_on_panic library to ensure that the
entire kernel is torn down if one of the processes panics. In turn
this should allow the feedback stack to collect a real stacktrace from
the crashing process.

Original-Fixed: b/333891263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028894
Original-Revision: ebf5cc45239381104aa2fcb7c98279b91d216f10
GitOrigin-RevId: bcb102c7cc2e3eb492e3ffa736e1f1e743715f95
Change-Id: I8dc6aa0eb746f597a628c976d84f7608c721346e