[roll] Roll fuchsia [pkgdir] Refactor tests to work for CTF
A breaking change caused a problem with this test's use of Dictionary
methods, which means it will fail when run against the frozen version at
F19.

We stabilized the Namespace/Create method which can take in the
Dictionary and hide the implementation details, but unfortunately we do
not yet support extending namespaces with directories that are not
readonly. In this case, we need the directory to support EXECUTABLE
rights.

For now, we are changing this test to pass in a server_end for the
directory in its input realm arguments and return an empty dictionary
that is ignored. In the future we will refactor this test to support
getting its output from the dictionary directly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034332
Original-Revision: 44fa6acd673b6fed98fc505677180f685039956f
GitOrigin-RevId: 6e737535190f68d324adf0debc8b47cb9bd112ea
Change-Id: I768be5aa23429304e57ca9aa03f140841d0a0141