[roll] Roll fuchsia [rust] Update fuchsia-zircon-types for release
Removes a dependency on static-assertions in favor of a short custom
macro. Switches zerocopy to an optional dependency so it can be enabled
and disabled. Updates the crate manifest with relevant information and
bumps the version to a new alpha.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1035092
Original-Revision: 7a281081c8f9aa157814d9c1d4e1bce76e15ce12
GitOrigin-RevId: 94f9c3f439c959fc3e59d459d1c9756887923890
Change-Id: I61aecd856cf9132cb01cb3c8a64c9b15a2d3d006