[roll] Roll infra/recipes Replace remaining Path.join() calls
Use Path.joinpath() or the / operator instead of Path.join().
Eventually, config_types.Path will be replaced with pathlib.Path, which
doesn't have a join() method. In general, if the substitution was simple
the / operator was used. Otherwise, '.join(' was replaced with
'.joinpath('.

Some calls remain that appear to be Path.join(str) calls but are
actually str.join(str) calls, at least in recipe tests. I suspect those
are unintended, but mostly left them as-is for now.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1035772
Original-Revision: dff30db5a2272ebe944356892bf92a7ca2992025
GitOrigin-RevId: 50e9169ffd17964d8fa814baef1ccd69280f673e
Change-Id: Ie7b13b12490b513d04be39eda4174bb2f2d1057f