[roll] Roll fuchsia [starnix] Add tests for various `brk()` edge-cases
Verify that pre-existing mappings prevent `brk()` being grown.
Verify that mappings over the program break do not prevent it
being reduced.
Verify that the program break address is the first byte after the
program break, not part of it, such that if the address is aligned
then the page containing it is not accessible.
Verify that shrinking to an aligned break, and re-growing,
correctly allocates fresh program break pages.
Verify that mappings over the existing break region do not prevent
the break from being extended, and are not overwritten.

Original-Bug: 334365844, 333435783, 333450691
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034794
Original-Revision: d21a1734f1bffcc2bf83e79c59a54aed32a27623
GitOrigin-RevId: 758e283eedfdf87c504fb72bf0edaff2b0e33905
Change-Id: I91e5eb9ea9eac4403d8a8b7d3603e0c79f3ffeae