[roll] Roll fuchsia [debugger] Make PC relative disassembly easier
Before this change you'd have to know what architecture you were on and
that you had to specify "*" on the disassemble command to tell it to do
some math to the PC before fetching the memory. Now just putting a '+'
or '-' in front of a raw number will be all it takes to get a PC
relative offset.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1022000
Original-Revision: a13b0338e1b49461fc2534d156483430b284ab38
GitOrigin-RevId: cb895690947f81fd3d7753d3a843872cd9731b1b
Change-Id: I9ec01d66f5737a350cb6ed24730c3fcf64672656