[roll] Roll fuchsia [devicetree] Preserve the devicetree node name
For use in driver framework, the devicetree node name was modified to
replace some of the special characters not supported by fdf. But the
name of the node is crucial for visitors while filtering out nodes that
they need to act upon. For this purpose, this CL adds a new name field
fdf_name() that holds the modified name of the node and the original
name is stored in the name() field as it is.

Test: *visitor-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1035278
Original-Revision: 650073f588906fea36f71fd8dcd7f73cdcfe2921
GitOrigin-RevId: 2a48e99790326e840bef536c326d25dc98007632
Change-Id: I949931617a155ddff23da3eae719855f0abec63c