[roll] Roll fuchsia [aml-power] Remove dependency on GetDeviceInfo
In preparation to adding devicetree support removing the dependency on
GetDeviceInfo by using a generic constructor. The check for
availability of properties and protocols is already done as part of
device-enumeration tests and is rightly setup by the respective board driver.

Hence consolidated the power domain information into a single argument
for the ctor. Also removed some of the unused code related to luis.

For consistency of mapping, all platforms that have A53 cores are mapped
to little domain and A73 cores are mapped to big domain.

Original-Bug: 333925701
Test: aml-meson-power-test, device enumeration tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027092
Original-Revision: cc84ee948a8970ebdeeffff60fa59c8a074d6144
GitOrigin-RevId: c9a5110d2615361e20dbca44662ade5e82ce6d00
Change-Id: Ib8d617f44fb7e0f7bf2a4c6f4d9907bd87276a5e