[roll] Roll fuchsia [versioning] Redefine FUCHSIA_HEAD for C++.
The "HEAD" API level is being redefined in this RFC: https://fxrev.dev/1009253

This CL updates the value of FUCHSIA_HEAD when used in a clang-y
context. Analogous changes for packages and FIDL soon to come.

I _think_ I got everything? But unfortunately 2^32-1 is not the easiest
number to search for...

Original-Bug: 335473877
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030293
Original-Revision: d2cb8360616b69d4cffd690e07f70f0c1643f290
GitOrigin-RevId: e8cab94f94fd4f1d99c18bfcc1e71e45429c52ff
Change-Id: I2cb79b3bff8728dd7bf486fd175ef6e6b964c79d