[roll] Roll fuchsia [lib/routing] Ensure dictionaries are not supported in legacy
We previously had a check in legacy routing that threw an error if a
component tried to offer or expose a protocol from a dictionary to
another component. However, it didn't fire unless the dictionary was
being used directly. If the protocol flows from a component that is
multiple hops away, it would not work.

Fix that issue, by throwing a dictionary not supported error whenever
a protocol is sourced from a dictionary, instead of a NOT_FOUND.

Original-Bug: 326086827
Test: Added unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036113
Original-Revision: 0947d7a2304d3ddc8d95dd5c02c5d41dc9847741
GitOrigin-RevId: 02e4beb9cc2bdd9b1b410c1bef38ca897535ccfd
Change-Id: I24d087faa3a1cefa837498e9fc609ac9188ac2b7