[roll] Roll fuchsia [ffx][package] Remove package plugin from ffx builtins
Removes ffx-package from being built in to the main ffx binary, allowing
it to be run from the SDK instead.

This does not change how users invoke the tool, only how it executes
when `ffx package` is invoked, but makes it so that both the tool and
ffx can build faster and with fewer resources making it easier to
iterate on them.

Original-Bug: b/329326263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916261
Original-Revision: 8c6571163678ffba3ac50934d8848a569c8f93c2
GitOrigin-RevId: fff266c14d83f7f5270ebff446a238e18b344a84
Change-Id: Ia501ec0e3b2ab08314b14e46858928399efa61fe