[roll] Roll fuchsia [vsock-sshd-host] Introduce a small vsock variant of sshd
This variant is not meant to provide the same facilities as the typical
tcp port 22 listening variant. Instead it provides the minimal
environment necessary to allow an ffx connection to establish over the
vsock.

This variant is meant to be used in configuration where a netstack is
not available and we are running in an emulator which supports vsock.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030780
Owners-Override: James Robinson <jamesr@google.com>
Original-Revision: d6cefd060fafdca498d1c14a40024471c40002b6
GitOrigin-RevId: d852b8fb45030dd00b18f31bf298e7371188bd0f
Change-Id: If6b725e9b46dfdf079a4665b8b1ad102a21b0d7c