[roll] Roll fuchsia [vim3][clk][dfv2] Replace Vim3 Clock Driver with DFv2
Vim3 previously used `aml-clock` as its clock driver. This is a DFv1
clock driver that is shared by all AMLogic platforms.

There is an ongoing effort to build Vim3 out-of-tree which would require
the Vim3 clock driver to be reimplemented using DFv2.

This CL migrates Vim3's clock functionality to an entirely new DFv2
driver that is exclusive to Vim3.

Limitations:
This driver does not implement the clock measure functionality of the
previous driver that allowed clients to measure the frequency of various
clocks in the system.

Testing:
 + Device Enumeration Test
   `fx test device-enumeration-test:vim3`
   Not only does this test that the new clock device is binding, it also
   tests that all downstream devices are able to find their clock
   dependencies and bind against them.
 + Unit Tests
   `fx test vim3_clk_driver_test`
    This is a new unit test that tests the functionality of this driver.
 + Spot Check
   I ran an additional spot check where I ran `k bench` and observed
   that the CPU was running at 2208000397 cycles per second which is
   consistent with the programmed frequency of 2.208GHz. Then I used
   `cpuctl` to set the core frequency to the lowest operating points using
   `cpuctl opp 000 10` and `cpuctl opp 001 6` and then ran `k bench again
   which yielded 999999800 cycles per second which at least demonstrates
   that the PLLs were able to be programmed by the CPU driver using the new
   DFv2 clock driver.

Original-Fixed: b/328302246

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1032061
Original-Revision: 0deb1f44eac9521ff9cb950bc4c6476e556ff9dd
GitOrigin-RevId: 1ba4db29a5b625490193fbea95b148afbe0ef722
Change-Id: I50680a30d048ed6e1cac89e17b7774f58f4736a0