[roll] Roll fuchsia [graphics] Change gpu_vendor_id to uint32 and expose as config
We can use hex in cml files to specify the gpu vendor ID, so it makes
sense to use it as a uint32, rather than have a string that needs to be
converted to a uint32. The value 0 is now treated as undefined, and an
arbitrary GPU will be selected with that value.

Also expose this value through the fuchsia.gpu.test.GpuVendorId config
capability.

These changes modify the interface of the code in the SDK, but no
clients were using the old config value, so it won't cause problems.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036112
Original-Revision: 9dc70d821183b578a2abd1d726519834f1377b18
GitOrigin-RevId: bea7f5946d33425598a061afcd9f9e63a490fa3b
Change-Id: I66e095e977e233f387eac97a7024848cff84b2d0