[roll] Roll fuchsia [power] Move domain creation to core driver
Power domains were created in the board driver using composite nodes.
There are a couple of issues with that mechanism: the logic will have to
be repeated in every board driver, this deviates from how gpio/i2c
children are created by the core drivers, this will be harder to create
when nested domain structure need to be created.

This CL updates the metadata passed to the power impl driver to include
domain information, which is forwarded to the power core. The power core
would internally create all the necessary domain nodes for relevant
power domain clients to bind to.
The motivation to make this change right now is it make the metadata and
node creation more streamlined for addition of devicetree support.

Original-Bug: 333925701
Test: device enumeration tests, power-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1027106
Original-Revision: dc6c2823091ee9eab28f45b2c0d667f1cc400a5b
GitOrigin-RevId: aeed79a59670b72a179169c478e86ba914baa1ce
Change-Id: I8ca313f84aaa422ac19d6d049a6d043ad4771d2a