[roll] Roll fuchsia [build][licenses] Generate full debug hints.
Ensure that all hints are recorded for a given CollectedLicense
instance, when the feature is enabled. Without this, which
hint ended up in the final file was semi-random since it
dependended on the GN graph traversal order, making
comparisons very difficult between changes.

With full hints, it becomes possible to understand much
more clearly why a certain license is or isn't in the
final output.

This generates larger SPDX files (e.g. the one for
@legacy_ninja_build_outputs goes from 4.5 MiB to 15 MiB), but
benchmarking shows:

- Minor speed difference in generation speed (1.9s -> 2.2s).
- No speed difference in license classification at all.

So this feature is always enabled for now.

+ generated_licenses_spdx_tool.py: Use WARNING instead of INFO
  as the default --log-level value for the script, to reduce
  build log spam. This will be important in future CLs because
  one SPDX file will be generated bazel_action() soon to
  populate the @gn_targets repository.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036772
Original-Revision: 62991722f382dfc4a457c9ec3cc8c7b22a15d459
GitOrigin-RevId: 1b2c356271b773e83ebfb39949262e6679d4a13e
Change-Id: Ib4000de594ea2c5fd9e51562860b96e0faad4090