[roll] Roll fuchsia [pkg-resolver] Do not cache blobfs results obtained outside the queue
Before adding the fetch of the meta.far blob to the queue,
pkg-resolver checks whether the meta.far is already resident. This
prevents blocking resolves of fully cached packages behind blob
fetches if the queue is already at its concurrency limit.

pkg-resolver was reusing the result of this check made outside the
queue in the fetch performed by the queue, which is incorrect,
because the queue may have already had a pending request to fetch
the blob, which would cause the result to be invalidated by the time
the queue started to process it, which would fail the fetch and
therefore the resolve.

With this change, pkg-resolver throws away the result and
rechecks for the blob's presence in the fetch performed by the queue.

Original-Fixed: 336589151
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1035239
Original-Revision: 905c18c5ba0899a26963cb9dc5fcf1343068c0cc
GitOrigin-RevId: 2588a7988cb11a03b71e6dbb63d03e3809ed247a
Change-Id: I64de9f292b6b5125b02e4185be2ab8c54c0c99d1