[roll] Roll fuchsia [graphics][vkproto] Change gpu_vendor_id to uint32 and expose as config
We can use hex in cml files to specify the gpu vendor ID, so it makes
sense to use it as a uint32, rather than have a string that needs to be
converted to a uint32. The value 0 is now treated as undefined, and an
arbitrary GPU will be selected with that value.

Also expose this value through the fuchsia.gpu.test.GpuVendorId config
capability.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037232
Original-Revision: df6fb560adf99b9ffafe8765911c3ca5df985d1a
GitOrigin-RevId: 99314c82f238571145485c9b586897db128e68cf
Change-Id: Ie3cb43b9b7522b666c4509b465d8261e7787723f