[roll] Roll fuchsia [testsharder] Fix to work with absolute package manifest paths.
If a test uses file-relative paths in its package manifest, the
LoadPackageManifests() function returns absolute paths if the base
manifest path it was given was an absolute path. However, testsharder
expected relative paths to the build dir and would thus prepend the
build dir to the path when trying to load subpackage manifests. Thus, we
should only prepend the build dir if the manifest path is not an
absolute path already.

Original-Bug: 337312770
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037587
Original-Revision: 4a2b99489c840da87500ec2ad0613c3689224514
GitOrigin-RevId: 8e0958fed16de6885bb0ebffa102815d9175ec35
Change-Id: Ic676c6a3661f3cf5142abe262aa605f02356d180