[roll] Roll fuchsia [debugger] Correctly break on rust await points
Adds additional processes to the line resolution path when resolving
line based breakpoints (or any other line-based symbol). There are two
steps:

First, line matches are cross referenced with the "line details" for the
match's address. The line details produce a range of addresses for the
match. Multiple matches may correspond to the same range of addresses.
From this set of matches for a particular address range, we can remove
all but the one with the lowest address, which is the "best". This step
is relatively cheap, and allows us to only perform the necessary symbol
lookups.

Once the line matches have been cross referenced with the line details,
we can compare the most specific code block that contains the PC for
each match. We again only need to keep the match with the lowest address
for each unique code block. The code blocks for Rust's generated code
around await points is both disjoint and in different code blocks, which
allows us to identify the correct locations such that a breakpoint would
actually be hit in a predictable location.

The above processing is a valid general strategy, and will result in us
choosing the correct address for setting breakpoints, but is
particularly important while working around the bug listed below, and
filed upstream at https://github.com/rust-lang/rust/issues/123341. When
the lowest address (which is typically what we choose as the "best" for
a range of addresses covering a particular line) is invalid, the
breakpoint will never be hit.

Adds a new prebuilt rust binary to the test_data for use in the
ModuleSymbols unit tests and does some small restructuring of the test
sources. There should be no changes to the C++ shared objects.

Original-Fixed: https://fxbug.dev/331475631
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1018577
Original-Revision: de5ac29abacfd060d5b66ba59c3a2e171e96fb31
GitOrigin-RevId: 24a9d1a630b5d689c5de1c47fc2f80c32adb783f
Change-Id: I0b55b3a37db22e2161104b0095ec98f95355671c