[roll] Roll fuchsia [assembly] Introduce a new AIB for ffx over vsock
This new AIB only makes sense for a product that targets virtual
machines and lack a network stack, such as microfuchsia. It's possible
to also enable it for use on other products that run on virtual machines
if we find a desire to avoid depending on the network stack, but that
will likely require another AIB with a core shard instead of a bootstrap
shard.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030782
Original-Revision: a41ac64693f660b8047d42624723a9434a8a4beb
GitOrigin-RevId: 6b45cef72f541525e107df18755023c38090b2eb
Change-Id: I64c9bb99262c6ba91f8aad680187fd118a0725e1