[roll] Roll fuchsia [pkg] Do not open blobs to check for presence
During resolves, after the meta.far is fetched, when checking to see
if a content blob or subpackage meta.far needs to be downloaded,
pkg-cache tries to open the blob. This is not correct, because on
c++blob READ connections to partially written blobs keep the blob
alive even after the writer closes its connection, which creates the
following race condition:

1. blob is partially written by a concurrent resolve
2. blob is opened to check for presence
3. concurrent resolve hits an error and retries the fetch, which
   attempts to open the blob for write, which fails

Local Storage has said that ReadDirents on /blob for both c++blob and
fxblob is guaranteed to return blobs if and only if they are fully
written and do not have outstanding deletion requests, this CL removes
the open double check and relies only on ReadDirents to determine
which blobs are missing.

Original-Fixed: 337104426
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037588
Original-Revision: 0f0c6fcd0d2f6f7d98f1de51616de274f277ede0
GitOrigin-RevId: f4a0b5819a704b2f709b49ea7c483b8b2374abcd
Change-Id: I69600e328009e6f272d5942a78c469241b52a9c2