[roll] Roll fuchsia [netstack3] Move GMP to TimerContext2
As one of our older modules, GMP was more prone to internal races so we
had to shift some code around here. This is a good showcase of the
nudging power of the new TimerContext, which forces the timer handling
to happen under locks and exposes the previous races.

In this CL we tackle IGMP, MLD, and the generic GMP module all at once
since they're closely related. Note that there's still a possible race
in these modules when sending reports out of locks, that's left for a
future change.

Note that we've removed the assertions for extra groups on
enabling/disabling IPv6 on a device. Looking inside installed timers for
the MLD module is probably overkill there and the flow is already
covered by local tests.

Original-Bug: 42083407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037936
Original-Revision: 69278fc3cdab7b0eb6620f77a18695200117fed1
GitOrigin-RevId: 440f34b36c84172f445447a923bccf43afc3a749
Change-Id: I74b4cfe3b9df485070b5ed19dcc1fa9b9a2e8edb