[roll] Roll fuchsia [netstack3] Move DAD to TimerContext2
Note that DAD is still not free from races with timers because the
timers are carrying an address by copy instead of a weak reference. That
part will be solved in a future change, when weak address IDs are
available.

Original-Bug: 42083407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1038134
Original-Revision: fed75ac7ddcda82558e57d1247f5471f17ac962e
GitOrigin-RevId: b3f2464918871b58f9102a6832c16de56fcd74e8
Change-Id: Ie1248fe91fdec7ee4337ca9e8eda76e5ac46d14b