[roll] Roll fuchsia [debugger] Automatically add source-map for rust stdlib
The rust stdlib sources are distributed as prebuilts, but do not seem to
adjust the DW_AT_comp_dir appropriately.

We can work around this for now with a default source-map setting while
the rust toolchain build is corrected.

Modifies the verbs_settings_unittest to not be dependent on path
resolutions for the new default value.

Original-Bug: https://fxbug.dev/336868560
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036353
Original-Revision: e2d858c6505023b5873e56cb28eea795b71eba42
GitOrigin-RevId: 310c688e985ba5d7f8e54da1c2cd38b0a1929d89
Change-Id: Id1c846f3b4846fad39dd4305b1170df20cf5ce4c