[roll] Roll fuchsia [dhcpv6/client] Add configurable DUID over FIDL
Add the ability to configure the DHCPv6 client's DUID over FIDL. In
particular, this allows using the DUID-LL and DUID-LLT methods of
generating a DUID using the link-layer address of a device, which
allows the DUID to be the same across restart of the client and
across reboots.

Netcfg has been changed to use DUID-LL with the MAC address of the
interface the DHCPv6 client is started on. This is good enough for
now, but the plan is eventually to use DUID-LLT and persist the value
to disk.

Tests were changed where needed. Also added a unit test for
constructing the DUID from FIDL types as the usage of `ArrayVec`
introduces panic opportunities.

Original-Fixed: 330190503
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028896
Original-Revision: fa5ec9113ae57b7a482262c9f7ec25478d3ddca0
GitOrigin-RevId: 8a8f6a584cd2169bc85fcd71405db7d3eb9e67c2
Change-Id: Id9680cad83fc7e5e721058e25dd6b7cb02041bbb