[roll] Roll fuchsia [graphics] Standardize package/component names for OOT tests
For tests that will soon be OOT, standardize on using hyphens in the
package name, having the package name end in "-test", and using a
"-test-child.cm" suffix for the component.

Also rename the cml files to match the .cm filename, to help when
codesearching.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037635
Original-Revision: 4affa52bc89f6d947e9637b6286840e5eae14e4c
GitOrigin-RevId: 258b5dbcdf91fd4bd7116783df99350ef5531aa8
Change-Id: I1961223da07e5495c1e567f3eebc7e2eae8f986a