[roll] Roll fuchsia [starnix] send flatland frames after GP connection
This change merges the `start_flatland_presentation_loop` and
`send_view_to_graphical_presenter` threads into a single, ordered set of
operations in which we (1) first connect to GraphicalPresenter and
PresentView to the server, then (2) start sending frames to Scenic Flatland.

Without this change, it's possible to start sending frames to Scenic
before it is ready, e.g. if Scene Manager has not set a root view yet,
causing frames to be dropped. Since we know that when PresentView is done,
Scenic and Scene Manager will both be ready, this is a better signal
after which we can start sending frames that we know will not get dropped.

Original-Bug: 333115991
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036081
Original-Revision: 92ce11fe994e695d3405be809197f01f45d133ba
GitOrigin-RevId: 9c63080c3e08fe10f0b806e7063f874231952728
Change-Id: I327b3b9647bf29a8ecccad998b43118d450e9fed