[roll] Roll fuchsia [netstack3] Extract FakeTimerCtx to netstack3_base
That allows the filtering crate to use the fakes as well.

Note that this extraction becomes a lot more permissive in terms of
`pub` symbols and `#[cfg(test)]` guards than what existed in core.
That's the price to pay for crating.

Original-Bug: 42086491
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1038974
Original-Revision: 16bb21d26e2d9332cece1fc80d99d48f982929c1
GitOrigin-RevId: f0f55b9add78b2a2b7488ed4f56d8f629506365f
Change-Id: Ib2a712cb909a3b1c9c761bebc0c1c51a5f64ebc6