[roll] Roll fuchsia [display][coordinator] Rework C-style APIs for getting display info.
This CL replaces APIs in DisplayInfo and Controller that use
out-parameters and C-style strings with APIs that use C++ types. This
highlights the lifecycle and synchronization subtleties in the Client
code that uses the APIs. The intent is to make it easier to reason about
future changes to DisplayInfo representation and to its processing by
Controller and Client code.

While at it, this CL clarifies the DisplayInfo copy and move support.

Original-Bug: 336394440
Test: fx test display-coordinator-unittest
Test: Booted core.qemu-x64 using ffx emu start --engine qemu
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1037116
Original-Revision: a2225042cb2e42ad5421090a4615038d5ac87356
GitOrigin-RevId: 0b4922b2e7105807168d75083500f8857c046c01
Change-Id: Iecde0da13572396c9fadd43ecb8b13b4edba6a1a