[roll] Roll fuchsia [fidl] Optimize the client waker
On Fuchsia, we can read the channel when we are woken and then only open
the task that needs to be woken rather than all of them.

On host, we can't do that because there are reentrancy issues, but we
can still avoid allocating a combined waker.

This is a reland. I fixed the reason for the most recent revert in
https://fxrev.dev/1028132.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030892
Original-Revision: e10ae0cc7e33abd43e7751c502761711ef8fce1e
GitOrigin-RevId: 595efd7304829df8706700c6935cd7d82b86d449
Change-Id: Ibe38cb789039da6151ca86ad006b9a28a26979a2