[roll] Roll fuchsia [netstack3] Implement TPROXY for UDP sockets
Implement the TransparentProxy filter action to transparently redirect
packets to a locally bound socket, for UDP sockets. This change does not
wire up transparent delivery for TCP sockets as it is not currently
needed. Packets are only proxied to a socket for which the
IP_TRANSPARENT option is enabled. Callers can use the IP_RECVORIGDSTADDR
receive control message to observe the original destination of proxied
packets.

Original-Bug: 337009139
Original-Fixed: 331819771
Test: tproxy-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034292
Original-Revision: 235395d1c580ebea09bb4965ab1c854914b64886
GitOrigin-RevId: 4f258344e0473e8900e2d7e3a4b22baf98de2912
Change-Id: I483dea581ce42851fe8fe87267bf93779ffbc885