[roll] Roll fuchsia [fxfs] Extend profile timeout in delete_profile test.
Previously the profile was run for zero seconds before being explicitly
stopped. Checks prior the explicit stop were relying on the profile to
have not yet completed, so this was obviously a race.

Extended the timeout to 10 minutes, which is longer than the timeout
that infra will grant it anyways, so it must always wait for the
explicit stop.

Original-Bug: b/327688777
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039413
Original-Revision: 270fe851979ec905876a05e689246471527a8571
GitOrigin-RevId: 42b67c93952ed2e7e9ca8241c980cdd5960351f9
Change-Id: I83548cd549990792c6f674c00ddbd26b7c36dfcb